refactor(resolve): split per-language patterns into separate files
Future #44 resolver slices no longer conflict on refs.go +
dynamic_patterns_test.go — each language has its own
dynamic_patterns_<lang>.go + _test.go file with self-registering
init() function.

Refs #44, Refs #1022 (closed v1 attempt due to rebase pain)
