+++ internal/resolve/dynamic_patterns_ruby.go
@@ -1,0 +1,207 @@
+package resolve
+
+import "regexp"
+
+// rubyDynamicPatterns are per-language patterns for Ruby.
+// Registered via init() into dynamicPatternsByLang.
+var rubyDynamicPatterns = []*regexp.Regexp{
+	// Bare-identifier forms: per-language extractors (Ruby, etc.)
+	// emit only the leaf callee identifier so the parens-requiring
+	// patterns below never match a real stub. Reflective Ruby method
+	// names are unique enough to be safe as bare-name anchors
+	// (issue #90).
+	regexp.MustCompile(`^send$`),
+	regexp.MustCompile(`^public_send$`),
+	regexp.MustCompile(`^__send__$`),
+	regexp.MustCompile(`^define_method$`),
+	regexp.MustCompile(`^instance_eval$`),
+	regexp.MustCompile(`^class_eval$`),
+	regexp.MustCompile(`^.*\.send\(`),        // obj.send(:name) — Ruby ONLY
+	regexp.MustCompile(`^send\(`),            // bare send(:name)
+	regexp.MustCompile(`^.*\.public_send\(`), // obj.public_send(:name)
+	regexp.MustCompile(`^public_send\(`),
+	regexp.MustCompile(`^.*\.__send__\(`),  // obj.__send__(:name)
+	regexp.MustCompile(`^method_missing$`), // ruby method_missing hook
+	regexp.MustCompile(`^.*\.method_missing\(`),
+	regexp.MustCompile(`^define_method\(`), // metaprogramming
+	regexp.MustCompile(`^.*\.define_method\(`),
+	regexp.MustCompile(`^.*\.instance_eval\(`),
+	regexp.MustCompile(`^.*\.class_eval\(`),
+
+	// Rails ActionController / ActionDispatch DSL methods (issue
+	// #107). These are method_missing-driven and routing/render
+	// helpers — bare-name calls in Rails controllers and views the
+	// resolver can't bind statically. Classify as Dynamic so they
+	// don't pollute bug-extractor (rails-realworld 38.93%, sidekiq
+	// 29.83% pre-fix). Names are Rails-unique enough not to collide
+	// with the generic Object/Kernel allowlist below.
+	regexp.MustCompile(`^render$`),
+	regexp.MustCompile(`^permit$`),
+	regexp.MustCompile(`^require$`),
+	regexp.MustCompile(`^redirect_to$`),
+	regexp.MustCompile(`^respond_to$`),
+	regexp.MustCompile(`^before_action$`),
+	regexp.MustCompile(`^skip_before_action$`),
+	regexp.MustCompile(`^after_action$`),
+	regexp.MustCompile(`^around_action$`),
+	regexp.MustCompile(`^helper_method$`),
+	regexp.MustCompile(`^params$`),
+	regexp.MustCompile(`^session$`),
+	regexp.MustCompile(`^flash$`),
+	regexp.MustCompile(`^cookies$`),
+	regexp.MustCompile(`^request$`),
+	regexp.MustCompile(`^response$`),
+	// ActiveRecord dynamic finders — `find_by_<attr>` /
+	// `find_or_create_by_<attr>` (with optional `!` bang variant)
+	// are method_missing-generated by AR at runtime.
+	regexp.MustCompile(`^find_by_\w+!?$`),
+	regexp.MustCompile(`^find_or_create_by_\w+!?$`),
+
+	// ActiveRecord query-builder methods (issue #107). Chained query
+	// DSL on AR relations — bare-name calls the resolver can't bind
+	// to a local entity. Multi-language collisions exist for
+	// `where`/`order`/etc., but the per-language gate (Ruby only)
+	// keeps them safely scoped. Generic collection ops (each / map /
+	// select / find / count / length / size) are deliberately
+	// EXCLUDED — they collide with user methods on any class.
+	regexp.MustCompile(`^order$`),
+	regexp.MustCompile(`^where$`),
+	regexp.MustCompile(`^joins$`),
+	regexp.MustCompile(`^includes$`),
+	regexp.MustCompile(`^eager_load$`),
+	regexp.MustCompile(`^preload$`),
+	regexp.MustCompile(`^pluck$`),
+	regexp.MustCompile(`^distinct$`),
+	regexp.MustCompile(`^group$`),
+	regexp.MustCompile(`^having$`),
+	regexp.MustCompile(`^limit$`),
+	regexp.MustCompile(`^offset$`),
+	regexp.MustCompile(`^scope$`),
+	regexp.MustCompile(`^belongs_to$`),
+	regexp.MustCompile(`^has_many$`),
+	regexp.MustCompile(`^has_one$`),
+	regexp.MustCompile(`^has_and_belongs_to_many$`),
+	regexp.MustCompile(`^validates$`),
+	regexp.MustCompile(`^validate$`),
+	regexp.MustCompile(`^before_save$`),
+	regexp.MustCompile(`^after_save$`),
+	regexp.MustCompile(`^before_create$`),
+	regexp.MustCompile(`^after_create$`),
+	regexp.MustCompile(`^before_destroy$`),
+	regexp.MustCompile(`^after_destroy$`),
+
+	// ActiveRecord migration DSL (issue #124). Rails migrations are
+	// method_missing-driven schema DSL — `t.string :name`,
+	// `create_table :users do |t|`, `add_index :users, :email`,
+	// `references :user`. The Ruby extractor strips the receiver and
+	// the resolver sees only the bare leaf identifier. These names
+	// are the dominant residue in rails-realworld bug-extractor
+	// after #143. Per-language gate (Ruby) keeps the column-type
+	// names (`string`, `integer`, `boolean`, `text`) safely scoped
+	// — they would collide trivially with user-method names in any
+	// other ecosystem.
+	regexp.MustCompile(`^create_table$`),
+	regexp.MustCompile(`^drop_table$`),
+	regexp.MustCompile(`^change_table$`),
+	regexp.MustCompile(`^rename_table$`),
+	regexp.MustCompile(`^add_column$`),
+	regexp.MustCompile(`^remove_column$`),
+	regexp.MustCompile(`^rename_column$`),
+	regexp.MustCompile(`^change_column$`),
+	regexp.MustCompile(`^add_index$`),
+	regexp.MustCompile(`^remove_index$`),
+	regexp.MustCompile(`^add_reference$`),
+	regexp.MustCompile(`^remove_reference$`),
+	regexp.MustCompile(`^add_foreign_key$`),
+	regexp.MustCompile(`^remove_foreign_key$`),
+	regexp.MustCompile(`^references$`),
+	regexp.MustCompile(`^timestamps$`),
+	regexp.MustCompile(`^string$`),
+	regexp.MustCompile(`^integer$`),
+	regexp.MustCompile(`^boolean$`),
+	regexp.MustCompile(`^text$`),
+	regexp.MustCompile(`^datetime$`),
+	regexp.MustCompile(`^date$`),
+	regexp.MustCompile(`^float$`),
+	regexp.MustCompile(`^decimal$`),
+	regexp.MustCompile(`^binary$`),
+	regexp.MustCompile(`^execute$`),
+
+	// Rails ActionPack / ActionDispatch / ActiveSupport internals
+	// (issue #448). Rails framework DSL exposed to controllers,
+	// routes and initializers — method_missing-generated or
+	// class-macro driven. The Ruby extractor strips the receiver
+	// and the resolver sees only the bare leaf identifier
+	// (`Rails.application.routes.draw { resources :users }` →
+	// `resources`/`draw`, `class_attribute :foo` → `class_attribute`,
+	// `Rails.application.config.middleware.insert_before(...)` →
+	// `insert_before`). Classify as Dynamic so they don't pollute
+	// bug-extractor (rails-actionpack 20.02% pre-fix).
+	//
+	// Conservative selection (lessons from #94 / #107):
+	// generic English verbs/accessors that exist as user-method
+	// names on any class in any language (`to`, `as`, `via`,
+	// `format`, `defaults`, `action`, `match`, `member`,
+	// `collection`, `nested`, `included`, `extended`, `inherited`,
+	// `concern` outside the routing-DSL position, `swap`) are
+	// deliberately EXCLUDED — the per-language Ruby gate alone is
+	// not strong enough to keep them safe. The names below are
+	// Rails-idiomatic enough that the Ruby gate is sufficient
+	// protection against shadowing user methods in other
+	// ecosystems (Go HTTP `get`/`post`, JS `get`/`post`, etc.).
+	//
+	// Categories:
+	//   - Routing DSL (ActionDispatch::Routing::Mapper).
+	//   - ActionController DSL macros (helper / layout / CSRF).
+	//   - ActiveSupport class macros and callbacks.
+	//   - ActionDispatch middleware-stack DSL.
+	// Already covered above by the issue #107 batch: `scope`,
+	// `helper_method`, `params`, `session`, `flash`, `cookies`,
+	// `request`, `response`, `respond_to`, `render`, `redirect_to`,
+	// `before_action`/`after_action`/`around_action`/
+	// `skip_before_action`.
+	// Routing DSL.
+	regexp.MustCompile(`^resources$`),
+	regexp.MustCompile(`^resource$`),
+	regexp.MustCompile(`^namespace$`),
+	regexp.MustCompile(`^constraints$`),
+	regexp.MustCompile(`^concern$`),
+	regexp.MustCompile(`^concerns$`),
+	regexp.MustCompile(`^mount$`),
+	regexp.MustCompile(`^get$`),
+	regexp.MustCompile(`^post$`),
+	regexp.MustCompile(`^put$`),
+	regexp.MustCompile(`^patch$`),
+	regexp.MustCompile(`^delete$`),
+	regexp.MustCompile(`^root$`),
+	regexp.MustCompile(`^direct$`),
+	regexp.MustCompile(`^resolve$`),
+	regexp.MustCompile(`^controller$`),
+	// ActionController DSL macros.
+	regexp.MustCompile(`^helper$`),
+	regexp.MustCompile(`^layout$`),
+	regexp.MustCompile(`^protect_from_forgery$`),
+	regexp.MustCompile(`^skip_authorization_check$`),
+	regexp.MustCompile(`^verify_authenticity_token$`),
+	regexp.MustCompile(`^respond_with$`),
+	regexp.MustCompile(`^headers$`),
+	// ActiveSupport class macros and callbacks.
+	regexp.MustCompile(`^prepended$`),
+	regexp.MustCompile(`^class_attribute$`),
+	regexp.MustCompile(`^mattr_accessor$`),
+	regexp.MustCompile(`^mattr_reader$`),
+	regexp.MustCompile(`^mattr_writer$`),
+	regexp.MustCompile(`^cattr_accessor$`),
+	regexp.MustCompile(`^define_callbacks$`),
+	regexp.MustCompile(`^set_callback$`),
+	regexp.MustCompile(`^skip_callback$`),
+	// ActionDispatch middleware-stack DSL.
+	regexp.MustCompile(`^add_middleware$`),
+	regexp.MustCompile(`^delete_middleware$`),
+	regexp.MustCompile(`^insert_before$`),
+	regexp.MustCompile(`^insert_after$`),
+}
+
+func init() {
+	dynamicPatternsByLang["ruby"] = rubyDynamicPatterns
+}
