+++ internal/resolve/refs.go
@@ -222,2279 +222,24 @@ var AllDispositions = []Disposition{
 // "java" (also "kotlin", "scala", "jvm"). Unknown / empty languages fall
 // back to crossLangDynamicPatterns only.
 var (
-	pythonDynamicPatterns = []*regexp.Regexp{
-		// Issue #432 — Python relative-import targets. The Python extractor
-		// preserves the leading dot for `from .compat import urlparse` and
-		// `from ..views import View` (extractor.go:653-655); the resolver
-		// then sees a bare ToID like `.compat.urlparse` or `..views.View`.
-		// One SCOPE.Component placeholder is emitted per importing file
-		// for the same module path, so bare-name lookup is ambiguous in
-		// any project where two or more files share the relative-import
-		// path — driving the edge to bug-resolver despite the placeholder
-		// being bookkeeping rather than the imported symbol's source.
-		// A leading-dot bare path is unambiguously a relative-import
-		// reference; route to Dynamic, mirroring the precedent for
-		// `scope:component:import:local:` (isHeuristicScopeStub) which
-		// the cross-language imports extractor emits for the same shape.
-		regexp.MustCompile(`^\.+[\w.]*$`),
-		// Bare-identifier forms: per-language extractors emit only the
-		// leaf callee identifier (e.g. ToID="getattr") for `getattr(...)`
-		// call sites. Without bare-name anchors none of the parens-
-		// requiring patterns below ever match real stubs (issue #90).
-		regexp.MustCompile(`^getattr$`),
-		regexp.MustCompile(`^setattr$`),
-		regexp.MustCompile(`^hasattr$`),
-		regexp.MustCompile(`^delattr$`),
-		// Wave-4 (Python) — `super().<method>(...)` chains. The Python
-		// extractor strips the `super()` receiver to a literal `super`
-		// segment, yielding stubs like `super.render`, `super.__init__`,
-		// `super.to_info_dict`. The dispatch target is the MRO-resolved
-		// parent method which depends on the (often-external) base
-		// class — overwhelmingly Dynamic-by-design, not an extractor bug.
-		regexp.MustCompile(`^super\.[A-Za-z_][A-Za-z0-9_]*`),
-		regexp.MustCompile(`^eval$`),
-		regexp.MustCompile(`^exec$`),
-		regexp.MustCompile(`^compile$`),
-		regexp.MustCompile(`^__import__$`),
-		regexp.MustCompile(`^hasattr\(`),              // hasattr(obj, name)
-		regexp.MustCompile(`^delattr\(`),              // delattr(obj, name)
-		regexp.MustCompile(`^compile\(`),              // compile(src, ...)
-		regexp.MustCompile(`^getattr\(`),              // getattr(obj, name)(...)
-		regexp.MustCompile(`^__getattr__$`),           // __getattr__ magic name
-		regexp.MustCompile(`^.*\.__getattr__\(`),      // obj.__getattr__("name")
-		regexp.MustCompile(`^.*\.__getattribute__\(`), // obj.__getattribute__(...)
-		regexp.MustCompile(`^setattr\(`),              // setattr-driven dispatch
-		regexp.MustCompile(`^globals\(\)\[`),          // globals()[name](...)
-		regexp.MustCompile(`^locals\(\)\[`),           // locals()[name](...)
-		regexp.MustCompile(`^vars\(\)\[`),             // vars()[name](...)
-		regexp.MustCompile(`^eval\(`),                 // eval(...)
-		regexp.MustCompile(`^exec\(`),                 // exec(...)
-		regexp.MustCompile(`^__import__\(`),           // __import__("modname")
-		regexp.MustCompile(`^importlib\.`),            // importlib.import_module / etc
-		regexp.MustCompile(`^functools\.partial\(`),   // functools.partial(...)
-		regexp.MustCompile(`^functools\.partialmethod\(`),
-		regexp.MustCompile(`^functools\.reduce\(`),
-		regexp.MustCompile(`^operator\.methodcaller\(`), // operator.methodcaller("name")
-		regexp.MustCompile(`^operator\.attrgetter\(`),   // operator.attrgetter(...)
-		regexp.MustCompile(`^operator\.itemgetter\(`),   // operator.itemgetter(...)
-		regexp.MustCompile(`^os\.environ\[`),            // env-driven (Python)
-		regexp.MustCompile(`^os\.getenv\(`),             // env-driven (Python)
-		// dispatch via dict/list subscript: handlers[key](...), funcs["x"](...).
-		// Anchored "<ident>[...](...)" so we don't bite plain attribute access.
-		regexp.MustCompile(`^[A-Za-z_][A-Za-z0-9_.]*\[[^\]]+\]\(`),
-
-		// Flask app-factory + decorator DSL (issue #420). Flask exposes
-		// its routing / lifecycle / CLI surface as decorator and method
-		// calls on `app = Flask(__name__)`, blueprints (`bp = Blueprint(...)`),
-		// and `bp.cli` AppGroup instances. The Python extractor strips
-		// the receiver and the resolver sees only the bare leaf
-		// identifier (e.g. `@app.route("/")` → ToID="route"). Without a
-		// per-language anchor those land in bug-extractor and drove
-		// flask + flask-realworld bug-rate to 43.93% / 43.63%. The
-		// per-language gate (Python only) keeps these names from
-		// shadowing user methods in other ecosystems — `route`, `errorhandler`,
-		// `before_request` etc. would collide trivially in Go / JS / Ruby.
-		//
-		// Mirrors the Rails ActionController DSL approach from #107.
-		regexp.MustCompile(`^route$`),                   // @app.route(...) / @bp.route(...)
-		regexp.MustCompile(`^add_url_rule$`),            // app.add_url_rule(...)
-		regexp.MustCompile(`^register_blueprint$`),      // app.register_blueprint(bp)
-		regexp.MustCompile(`^before_request$`),          // @app.before_request
-		regexp.MustCompile(`^before_first_request$`),    // @app.before_first_request (legacy)
-		regexp.MustCompile(`^after_request$`),           // @app.after_request
-		regexp.MustCompile(`^teardown_request$`),        // @app.teardown_request
-		regexp.MustCompile(`^teardown_appcontext$`),     // @app.teardown_appcontext
-		regexp.MustCompile(`^errorhandler$`),            // @app.errorhandler(404)
-		regexp.MustCompile(`^register_error_handler$`),  // app.register_error_handler(...)
-		regexp.MustCompile(`^shell_context_processor$`), // @app.shell_context_processor
-		regexp.MustCompile(`^context_processor$`),       // @app.context_processor
-		regexp.MustCompile(`^template_filter$`),         // @app.template_filter(...)
-		regexp.MustCompile(`^template_test$`),           // @app.template_test(...)
-		regexp.MustCompile(`^template_global$`),         // @app.template_global(...)
-		regexp.MustCompile(`^url_value_preprocessor$`),  // @app.url_value_preprocessor
-		regexp.MustCompile(`^url_defaults$`),            // @app.url_defaults
-		regexp.MustCompile(`^before_app_request$`),      // blueprint scoped variants
-		regexp.MustCompile(`^before_app_first_request$`),
-		regexp.MustCompile(`^after_app_request$`),
-		regexp.MustCompile(`^teardown_app_request$`),
-		regexp.MustCompile(`^app_errorhandler$`),
-		regexp.MustCompile(`^app_context_processor$`),
-		regexp.MustCompile(`^app_template_filter$`),
-		regexp.MustCompile(`^app_template_test$`),
-		regexp.MustCompile(`^app_template_global$`),
-		regexp.MustCompile(`^app_url_value_preprocessor$`),
-		regexp.MustCompile(`^app_url_defaults$`),
-		regexp.MustCompile(`^record$`),      // @bp.record (blueprint setup-state hook)
-		regexp.MustCompile(`^record_once$`), // @bp.record_once
-		// Flask CLI / click AppGroup decorator: `@bp.cli.command(...)` and
-		// `@app.cli.command(...)`. Extractor leaf is "command".
-		regexp.MustCompile(`^command$`),
-
-		// click decorator + helper DSL (issue #423). click is the
-		// dominant Python CLI framework and its decorators (`@click.command()`,
-		// `@click.group()`, `@click.option('--foo')`, `@click.argument('name')`,
-		// `@click.pass_context`) plus helper functions (`click.echo`,
-		// `click.prompt`, `click.confirm`, `click.style`, ...) arrive at
-		// the resolver as bare leaf identifiers after the Python extractor
-		// strips the `click.` receiver. Without anchors here every
-		// decorator/helper call site inflates bug-extractor (python/click
-		// 32.60% pre-fix). Names follow the Rails (#107) and Flask (#420)
-		// precedent: collision with user methods is accepted because the
-		// per-language gate (Python only) keeps them safely scoped, and
-		// Dynamic is the appropriate "we know it's framework dispatch we
-		// can't statically resolve" bucket. click constants (`STRING`,
-		// `INT`, `FLOAT`, `BOOL`, `UUID`) and class types (`Path`,
-		// `Choice`, `IntRange`, `FloatRange`, `Tuple`, `File`,
-		// `make_pass_decorator`) are deliberately EXCLUDED here — they
-		// arrive as `ext:click.<name>` stubs and the external allowlist
-		// (click is allowlisted) classifies them ExternalKnown without
-		// help from the dynamic catalog.
-		// NOTE: ^command$ already declared above by the Flask block (#420).
-		regexp.MustCompile(`^group$`),
-		regexp.MustCompile(`^option$`),
-		regexp.MustCompile(`^argument$`),
-		regexp.MustCompile(`^pass_context$`),
-		regexp.MustCompile(`^pass_obj$`),
-		regexp.MustCompile(`^pass_meta_key$`),
-		regexp.MustCompile(`^echo$`),
-		regexp.MustCompile(`^secho$`),
-		regexp.MustCompile(`^prompt$`),
-		regexp.MustCompile(`^confirm$`),
-		regexp.MustCompile(`^progressbar$`),
-		regexp.MustCompile(`^getchar$`),
-		regexp.MustCompile(`^pause$`),
-		regexp.MustCompile(`^clear$`),
-		regexp.MustCompile(`^style$`),
-		regexp.MustCompile(`^unstyle$`),
-		regexp.MustCompile(`^format_filename$`),
-		regexp.MustCompile(`^get_terminal_size$`),
-		regexp.MustCompile(`^launch$`),
-		regexp.MustCompile(`^edit$`),
-		regexp.MustCompile(`^get_app_dir$`),
-
-		// Flask extensions + Marshmallow + Flask-SQLAlchemy DSL (issue
-		// #446). Residual after #420 was Flask-SQLAlchemy column / type
-		// / relationship constructors on `db = SQLAlchemy()`, Flask-Login
-		// proxies (`current_user`, `@login_required`), Flask-WTF form
-		// methods (`form.validate_on_submit()`), Marshmallow schema field
-		// constructors (`fields.Str`, `fields.Nested`) and (de)serialization
-		// hooks (`@pre_load`, `Schema.dump`, `Schema.load`), and Flask's
-		// common response helpers (`jsonify`, `abort`, `send_file`). The
-		// Python extractor strips receivers like `db.`, `fields.`,
-		// `Schema.`, `form.`, `app.` and only the bare leaf identifier
-		// arrives at the resolver. Pre-fix this drove flask 41.32% and
-		// flask-realworld 43.47%. Per-language gate (Python only) keeps
-		// generic leaves (`add`, `delete`, `commit`, `session`, `query`,
-		// `dump`, `load`, `fields`, `String`, `Integer`) from shadowing
-		// user methods/types in other ecosystems. Within Python the
-		// collision trade is accepted: same precedent as Rails
-		// `render`/`session`/`params` (#107) and Flask `route`/`command`
-		// (#420) — Dynamic is the appropriate bucket for framework
-		// dispatch the resolver can't statically bind.
-		// Flask-SQLAlchemy
-		regexp.MustCompile(`^Column$`),
-		regexp.MustCompile(`^ForeignKey$`),
-		regexp.MustCompile(`^relationship$`),
-		regexp.MustCompile(`^backref$`),
-		regexp.MustCompile(`^Integer$`),
-		regexp.MustCompile(`^String$`),
-		regexp.MustCompile(`^Text$`),
-		regexp.MustCompile(`^Boolean$`),
-		regexp.MustCompile(`^DateTime$`),
-		regexp.MustCompile(`^Date$`),
-		regexp.MustCompile(`^Float$`),
-		regexp.MustCompile(`^Numeric$`),
-		regexp.MustCompile(`^init_app$`),
-		regexp.MustCompile(`^query$`),
-		regexp.MustCompile(`^query_property$`),
-		regexp.MustCompile(`^create_all$`),
-		regexp.MustCompile(`^drop_all$`),
-		regexp.MustCompile(`^session$`),
-		regexp.MustCompile(`^commit$`),
-		regexp.MustCompile(`^rollback$`),
-		regexp.MustCompile(`^flush$`),
-		regexp.MustCompile(`^add$`),
-		regexp.MustCompile(`^delete$`),
-		regexp.MustCompile(`^merge$`),
-		regexp.MustCompile(`^refresh$`),
-		// Flask-Login
-		regexp.MustCompile(`^current_user$`),
-		regexp.MustCompile(`^login_required$`),
-		regexp.MustCompile(`^login_user$`),
-		regexp.MustCompile(`^logout_user$`),
-		regexp.MustCompile(`^confirm_login$`),
-		// Flask-WTF
-		regexp.MustCompile(`^validate_on_submit$`),
-		regexp.MustCompile(`^populate_obj$`),
-		regexp.MustCompile(`^render_kw$`),
-		// Marshmallow — `Boolean` and `DateTime` overlap with the
-		// SQLAlchemy types above and are already covered.
-		regexp.MustCompile(`^fields$`),
-		regexp.MustCompile(`^Schema$`),
-		regexp.MustCompile(`^Str$`),
-		regexp.MustCompile(`^Int$`),
-		regexp.MustCompile(`^List$`),
-		regexp.MustCompile(`^Nested$`),
-		regexp.MustCompile(`^Method$`),
-		regexp.MustCompile(`^Function$`),
-		regexp.MustCompile(`^pre_load$`),
-		regexp.MustCompile(`^post_load$`),
-		regexp.MustCompile(`^pre_dump$`),
-		regexp.MustCompile(`^post_dump$`),
-		regexp.MustCompile(`^validates$`),
-		regexp.MustCompile(`^validates_schema$`),
-		regexp.MustCompile(`^dump$`),
-		regexp.MustCompile(`^load$`),
-		regexp.MustCompile(`^dumps$`),
-		regexp.MustCompile(`^loads$`),
-		// Flask common response helpers
-		regexp.MustCompile(`^jsonify$`),
-		regexp.MustCompile(`^make_response$`),
-		regexp.MustCompile(`^abort$`),
-		regexp.MustCompile(`^send_file$`),
-		regexp.MustCompile(`^send_from_directory$`),
-		regexp.MustCompile(`^stream_with_context$`),
-	}
-
-	goDynamicPatterns = []*regexp.Regexp{
-		regexp.MustCompile(`^reflect\.`),       // reflect.* (Call, ValueOf, MethodByName, ...)
-		regexp.MustCompile(`\.MethodByName\(`), // v.MethodByName("X").Call(...)
-		regexp.MustCompile(`\.FieldByName\(`),  // v.FieldByName("X")
-		regexp.MustCompile(`^plugin\.Open\(`),  // Go plugin loader
-		// Anchored: only `plugin.Lookup(` (or `<x>.plugin.Lookup(`) — bare
-		// `repo.Lookup(id)` / `cache.Lookup(...)` are NOT reflection.
-		regexp.MustCompile(`\bplugin\.Lookup\(`),
-	}
-
-	jsDynamicPatterns = []*regexp.Regexp{
-		regexp.MustCompile(`^Reflect\.`),      // Reflect.apply / Reflect.construct / Reflect.get
-		regexp.MustCompile(`^eval$`),          // bare eval (issue #95)
-		regexp.MustCompile(`^eval\(`),         // eval(src)
-		regexp.MustCompile(`^Function$`),      // bare Function constructor reference
-		regexp.MustCompile(`^Function\(`),     // Function(src)
-		regexp.MustCompile(`^new Function\(`), // new Function(src)
-		// Dynamic import / require: must NOT be a literal-string first arg —
-		// `require("fs")` and `import("./mod")` are statically resolvable.
-		regexp.MustCompile("^require\\([^\"'`)]"),
-		regexp.MustCompile("^import\\([^\"'`)]"),
-		regexp.MustCompile(`^process\.env\.`), // env-driven (JS)
-		// Wave-4 (TS framework) — relative-import paths. The JS/TS
-		// extractor emits IMPORTS edges with the literal module string
-		// as ToID (e.g. `./cart-context`, `../fragments/cart`, `.`,
-		// `..`, `../..`). Every importing file produces its own
-		// SCOPE.Component import entity for the same module string, so
-		// bare-name lookup is ambiguous and the edge drives to
-		// bug-resolver despite the placeholder being bookkeeping rather
-		// than the imported symbol's source. Mirrors the Python relative-
-		// import pattern above (#432) and the
-		// `scope:component:import:local:` heuristic-scope-stub branch
-		// (which the cross-language imports extractor emits for the same
-		// shape). Pattern matches `.`, `..`, and any leading-`./`/`../`
-		// path; anchored to avoid biting bare identifiers.
-		regexp.MustCompile(`^\.{1,3}$`),
-		regexp.MustCompile(`^\.{1,2}/`),
-		// TS path-mapped local imports — tsconfig.json `baseUrl` /
-		// `paths` lets imports look like `components/grid` or
-		// `lib/shopify/types`. These are intra-repo (no leading dot, no
-		// leading `@scope`, no `node:` prefix, no package dot-domain
-		// like `next.js`) and the extractor emits one SCOPE.Component
-		// per importing file, producing the same ambig-bare-no-hint
-		// disposition as relative paths. Restrict to multi-segment
-		// paths whose first segment is a common TS-monorepo source root
-		// (`src`, `app`, `lib`, `components`, `pages`, `hooks`,
-		// `utils`, `helpers`, `services`, `store`, `styles`, `types`,
-		// `config`, `constants`, `features`, `modules`, `domain`,
-		// `data`, `api`, `server`, `client`, `shared`, `core`,
-		// `common`, `models`, `views`, `controllers`, `middleware`,
-		// `tests`, `test`, `__tests__`, `__mocks__`, `routes`). The
-		// per-language gate (js/ts only) keeps these names from
-		// shadowing real go/python/etc. modules with the same prefix.
-		regexp.MustCompile(`^(src|app|lib|components|pages|hooks|utils|helpers|services|store|styles|types|config|constants|features|modules|domain|data|api|server|client|shared|core|common|models|views|controllers|middleware|tests|test|__tests__|__mocks__|routes)/`),
-		// JS reflective `Function.prototype.{bind,apply,call}` is real, but
-		// the bare `.bind(` / `.apply(` / `.call(` patterns collide with too
-		// many domain methods (DB driver `bind`, `discount.apply(order)`,
-		// `controller.call(...)`). Keep them out of the JS catalog; the
-		// extractors tag truly reflective uses (e.g. `Reflect.apply`) which
-		// the explicit `Reflect\.` pattern above already covers.
-
-		// Wave-7 (TS/JS React frontend, #519) — React useState setter
-		// destructure pattern. The JS extractor strips the receiver from
-		// destructured tuples (`const [v, setV] = useState(...)`), leaving
-		// bare `setV` callee names that the resolver cannot bind because
-		// the symbol is component-local and the extractor doesn't know its
-		// origin. The `set[A-Z]...` convention is universal in the React
-		// community (RFC + React docs) and the per-language gate
-		// (js/ts only) prevents collision with `setHeader` / `setCookie`
-		// style helpers in non-JS code. Names are bare leaf identifiers.
-		regexp.MustCompile(`^set[A-Z][A-Za-z0-9_]*$`),
-		// Promise chain methods — `then`, `catch`, `finally` — bare-name
-		// callees on the result of `await`-able / Promise-returning
-		// functions. The extractor emits the chained method as a bare
-		// identifier with the receiver stripped, and the receiver is a
-		// Promise value the resolver cannot model. `then` alone is the
-		// dominant residual in client-fixture-b. JS-only gate keeps these
-		// out of Ruby (`then` is a real method) / Go / Python collisions.
-		regexp.MustCompile(`^then$`),
-		regexp.MustCompile(`^catch$`),
-		regexp.MustCompile(`^finally$`),
-		// Wave-11 (TS/JS React frontend, ship-gate) — React handler-prop
-		// convention `onClose`, `onClick`, `onChange`, `onSubmit`,
-		// `onCancel`, `onConfirm`, `onSuccess`, `onError`, `onValueChange`,
-		// `onSelect`, `onFocus`, `onBlur`, etc. These are callable props
-		// passed from a parent component; the actual handler implementation
-		// lives in the parent and is bound at component-invocation time, so
-		// the call site (inside the child component) cannot statically bind
-		// it. Standard React convention (React docs + RFC). The per-language
-		// gate (js/ts only) keeps it from biting Python/Go/etc. where the
-		// `onX` callable-prop convention does not exist. client-fixture-b
-		// top bug-extractor sample after wave-10 was `onClearAll` /
-		// `onClose` confirming this is the dominant residual shape.
-		regexp.MustCompile(`^on[A-Z][A-Za-z0-9]*$`),
-		// Wave-13 (TS/JS React frontend, real-residue) — `handle*` and
-		// `after*` callable-prop / lifecycle-hook conventions. Mirrors
-		// the wave-11 `^on[A-Z]...` rule. React/JSX components routinely
-		// destructure callable props named `handleClientSelection`,
-		// `handleReloadData`, `handleSaveOnCell`, `afterSaveNote`,
-		// `afterCreateSuccess` from parent components or pass them as
-		// `useCallback` returns. The actual implementation lives in the
-		// parent (or in a higher-order hook) and is bound at component-
-		// invocation time, so the call site cannot statically bind it.
-		// `handleX` is universal React tutorial style (React docs +
-		// every state-of-the-art repo); `afterX` is form/lifecycle hook
-		// convention (antd Form `afterClose`, react-hook-form `afterSubmit`).
-		// Both names dominate cfb wave-12-FINAL bug-extractor + bug-resolver
-		// residues (`handleClientSelection`, `handleReloadData`,
-		// `afterSaveNote`, `afterSaveSuccess`, `afterCreateSuccess`).
-		// Same-file preference resolver (wave-9 Chain-fix A) fires BEFORE
-		// the dynamic pattern check via the hex-ID branch in
-		// classifyDispositionLang, so a same-file lifted handler entity
-		// still wins. Per-language gate (js/ts only) keeps these from
-		// shadowing `HandleXxx` Go method conventions / Python `handle_*`
-		// snake_case (covered by other patterns). Conservative scope:
-		// `get*` / `set*` / `load*` / `save*` / `create*` / `update*` /
-		// `delete*` / `fetch*` / `use*` / `submit*` / `cancel*` / `select*`
-		// / `reset*` / `toggle*` are deliberately EXCLUDED — those verbs
-		// shadow real user-defined entities (services, repos, mutators)
-		// far more aggressively than `handle*`/`after*`.
-		regexp.MustCompile(`^handle[A-Z][A-Za-z0-9]*$`),
-		regexp.MustCompile(`^after[A-Z][A-Za-z0-9]*$`),
-		// PLT #537 RN/Expo wave — React ref + responsive-design + RN
-		// notification idioms that the JS extractor strips to bare leaf
-		// identifiers. `current` is the universal React useRef property
-		// (`ref.current`); `isTablet` / `isMobile` / `isLandscape` /
-		// `isPortrait` are responsive-design flags returned by
-		// react-native-device-info / useResponsive hooks and consumed as
-		// bare destructured locals; `enqueue` is the RN notification
-		// queue method on the NotificationContext / Toast / Snackbar
-		// receivers (also antd's notification.enqueue). These names are
-		// indistinguishable from real user methods to the resolver
-		// after the receiver is stripped — same shape as the wave-11
-		// `^on[A-Z]` rule. JS-only gate keeps them out of Go / Python /
-		// Java where `current` is a real first-class symbol name.
-		regexp.MustCompile(`^current$`),
-		regexp.MustCompile(`^isTablet$`),
-		regexp.MustCompile(`^isMobile$`),
-		regexp.MustCompile(`^isLandscape$`),
-		regexp.MustCompile(`^isPortrait$`),
-		regexp.MustCompile(`^isDesktop$`),
-		regexp.MustCompile(`^enqueue$`),
-	}
-
-	rubyDynamicPatterns = []*regexp.Regexp{
-		// Bare-identifier forms: per-language extractors (Ruby, etc.)
-		// emit only the leaf callee identifier so the parens-requiring
-		// patterns below never match a real stub. Reflective Ruby method
-		// names are unique enough to be safe as bare-name anchors
-		// (issue #90).
-		regexp.MustCompile(`^send$`),
-		regexp.MustCompile(`^public_send$`),
-		regexp.MustCompile(`^__send__$`),
-		regexp.MustCompile(`^define_method$`),
-		regexp.MustCompile(`^instance_eval$`),
-		regexp.MustCompile(`^class_eval$`),
-		regexp.MustCompile(`^.*\.send\(`),        // obj.send(:name) — Ruby ONLY
-		regexp.MustCompile(`^send\(`),            // bare send(:name)
-		regexp.MustCompile(`^.*\.public_send\(`), // obj.public_send(:name)
-		regexp.MustCompile(`^public_send\(`),
-		regexp.MustCompile(`^.*\.__send__\(`),  // obj.__send__(:name)
-		regexp.MustCompile(`^method_missing$`), // ruby method_missing hook
-		regexp.MustCompile(`^.*\.method_missing\(`),
-		regexp.MustCompile(`^define_method\(`), // metaprogramming
-		regexp.MustCompile(`^.*\.define_method\(`),
-		regexp.MustCompile(`^.*\.instance_eval\(`),
-		regexp.MustCompile(`^.*\.class_eval\(`),
-
-		// Rails ActionController / ActionDispatch DSL methods (issue
-		// #107). These are method_missing-driven and routing/render
-		// helpers — bare-name calls in Rails controllers and views the
-		// resolver can't bind statically. Classify as Dynamic so they
-		// don't pollute bug-extractor (rails-realworld 38.93%, sidekiq
-		// 29.83% pre-fix). Names are Rails-unique enough not to collide
-		// with the generic Object/Kernel allowlist below.
-		regexp.MustCompile(`^render$`),
-		regexp.MustCompile(`^permit$`),
-		regexp.MustCompile(`^require$`),
-		regexp.MustCompile(`^redirect_to$`),
-		regexp.MustCompile(`^respond_to$`),
-		regexp.MustCompile(`^before_action$`),
-		regexp.MustCompile(`^skip_before_action$`),
-		regexp.MustCompile(`^after_action$`),
-		regexp.MustCompile(`^around_action$`),
-		regexp.MustCompile(`^helper_method$`),
-		regexp.MustCompile(`^params$`),
-		regexp.MustCompile(`^session$`),
-		regexp.MustCompile(`^flash$`),
-		regexp.MustCompile(`^cookies$`),
-		regexp.MustCompile(`^request$`),
-		regexp.MustCompile(`^response$`),
-		// ActiveRecord dynamic finders — `find_by_<attr>` /
-		// `find_or_create_by_<attr>` (with optional `!` bang variant)
-		// are method_missing-generated by AR at runtime.
-		regexp.MustCompile(`^find_by_\w+!?$`),
-		regexp.MustCompile(`^find_or_create_by_\w+!?$`),
-
-		// ActiveRecord query-builder methods (issue #107). Chained query
-		// DSL on AR relations — bare-name calls the resolver can't bind
-		// to a local entity. Multi-language collisions exist for
-		// `where`/`order`/etc., but the per-language gate (Ruby only)
-		// keeps them safely scoped. Generic collection ops (each / map /
-		// select / find / count / length / size) are deliberately
-		// EXCLUDED — they collide with user methods on any class.
-		regexp.MustCompile(`^order$`),
-		regexp.MustCompile(`^where$`),
-		regexp.MustCompile(`^joins$`),
-		regexp.MustCompile(`^includes$`),
-		regexp.MustCompile(`^eager_load$`),
-		regexp.MustCompile(`^preload$`),
-		regexp.MustCompile(`^pluck$`),
-		regexp.MustCompile(`^distinct$`),
-		regexp.MustCompile(`^group$`),
-		regexp.MustCompile(`^having$`),
-		regexp.MustCompile(`^limit$`),
-		regexp.MustCompile(`^offset$`),
-		regexp.MustCompile(`^scope$`),
-		regexp.MustCompile(`^belongs_to$`),
-		regexp.MustCompile(`^has_many$`),
-		regexp.MustCompile(`^has_one$`),
-		regexp.MustCompile(`^has_and_belongs_to_many$`),
-		regexp.MustCompile(`^validates$`),
-		regexp.MustCompile(`^validate$`),
-		regexp.MustCompile(`^before_save$`),
-		regexp.MustCompile(`^after_save$`),
-		regexp.MustCompile(`^before_create$`),
-		regexp.MustCompile(`^after_create$`),
-		regexp.MustCompile(`^before_destroy$`),
-		regexp.MustCompile(`^after_destroy$`),
-
-		// ActiveRecord migration DSL (issue #124). Rails migrations are
-		// method_missing-driven schema DSL — `t.string :name`,
-		// `create_table :users do |t|`, `add_index :users, :email`,
-		// `references :user`. The Ruby extractor strips the receiver and
-		// the resolver sees only the bare leaf identifier. These names
-		// are the dominant residue in rails-realworld bug-extractor
-		// after #143. Per-language gate (Ruby) keeps the column-type
-		// names (`string`, `integer`, `boolean`, `text`) safely scoped
-		// — they would collide trivially with user-method names in any
-		// other ecosystem.
-		regexp.MustCompile(`^create_table$`),
-		regexp.MustCompile(`^drop_table$`),
-		regexp.MustCompile(`^change_table$`),
-		regexp.MustCompile(`^rename_table$`),
-		regexp.MustCompile(`^add_column$`),
-		regexp.MustCompile(`^remove_column$`),
-		regexp.MustCompile(`^rename_column$`),
-		regexp.MustCompile(`^change_column$`),
-		regexp.MustCompile(`^add_index$`),
-		regexp.MustCompile(`^remove_index$`),
-		regexp.MustCompile(`^add_reference$`),
-		regexp.MustCompile(`^remove_reference$`),
-		regexp.MustCompile(`^add_foreign_key$`),
-		regexp.MustCompile(`^remove_foreign_key$`),
-		regexp.MustCompile(`^references$`),
-		regexp.MustCompile(`^timestamps$`),
-		regexp.MustCompile(`^string$`),
-		regexp.MustCompile(`^integer$`),
-		regexp.MustCompile(`^boolean$`),
-		regexp.MustCompile(`^text$`),
-		regexp.MustCompile(`^datetime$`),
-		regexp.MustCompile(`^date$`),
-		regexp.MustCompile(`^float$`),
-		regexp.MustCompile(`^decimal$`),
-		regexp.MustCompile(`^binary$`),
-		regexp.MustCompile(`^execute$`),
-
-		// Rails ActionPack / ActionDispatch / ActiveSupport internals
-		// (issue #448). Rails framework DSL exposed to controllers,
-		// routes and initializers — method_missing-generated or
-		// class-macro driven. The Ruby extractor strips the receiver
-		// and the resolver sees only the bare leaf identifier
-		// (`Rails.application.routes.draw { resources :users }` →
-		// `resources`/`draw`, `class_attribute :foo` → `class_attribute`,
-		// `Rails.application.config.middleware.insert_before(...)` →
-		// `insert_before`). Classify as Dynamic so they don't pollute
-		// bug-extractor (rails-actionpack 20.02% pre-fix).
-		//
-		// Conservative selection (lessons from #94 / #107):
-		// generic English verbs/accessors that exist as user-method
-		// names on any class in any language (`to`, `as`, `via`,
-		// `format`, `defaults`, `action`, `match`, `member`,
-		// `collection`, `nested`, `included`, `extended`, `inherited`,
-		// `concern` outside the routing-DSL position, `swap`) are
-		// deliberately EXCLUDED — the per-language Ruby gate alone is
-		// not strong enough to keep them safe. The names below are
-		// Rails-idiomatic enough that the Ruby gate is sufficient
-		// protection against shadowing user methods in other
-		// ecosystems (Go HTTP `get`/`post`, JS `get`/`post`, etc.).
-		//
-		// Categories:
-		//   - Routing DSL (ActionDispatch::Routing::Mapper).
-		//   - ActionController DSL macros (helper / layout / CSRF).
-		//   - ActiveSupport class macros and callbacks.
-		//   - ActionDispatch middleware-stack DSL.
-		// Already covered above by the issue #107 batch: `scope`,
-		// `helper_method`, `params`, `session`, `flash`, `cookies`,
-		// `request`, `response`, `respond_to`, `render`, `redirect_to`,
-		// `before_action`/`after_action`/`around_action`/
-		// `skip_before_action`.
-		// Routing DSL.
-		regexp.MustCompile(`^resources$`),
-		regexp.MustCompile(`^resource$`),
-		regexp.MustCompile(`^namespace$`),
-		regexp.MustCompile(`^constraints$`),
-		regexp.MustCompile(`^concern$`),
-		regexp.MustCompile(`^concerns$`),
-		regexp.MustCompile(`^mount$`),
-		regexp.MustCompile(`^get$`),
-		regexp.MustCompile(`^post$`),
-		regexp.MustCompile(`^put$`),
-		regexp.MustCompile(`^patch$`),
-		regexp.MustCompile(`^delete$`),
-		regexp.MustCompile(`^root$`),
-		regexp.MustCompile(`^direct$`),
-		regexp.MustCompile(`^resolve$`),
-		regexp.MustCompile(`^controller$`),
-		// ActionController DSL macros.
-		regexp.MustCompile(`^helper$`),
-		regexp.MustCompile(`^layout$`),
-		regexp.MustCompile(`^protect_from_forgery$`),
-		regexp.MustCompile(`^skip_authorization_check$`),
-		regexp.MustCompile(`^verify_authenticity_token$`),
-		regexp.MustCompile(`^respond_with$`),
-		regexp.MustCompile(`^headers$`),
-		// ActiveSupport class macros and callbacks.
-		regexp.MustCompile(`^prepended$`),
-		regexp.MustCompile(`^class_attribute$`),
-		regexp.MustCompile(`^mattr_accessor$`),
-		regexp.MustCompile(`^mattr_reader$`),
-		regexp.MustCompile(`^mattr_writer$`),
-		regexp.MustCompile(`^cattr_accessor$`),
-		regexp.MustCompile(`^define_callbacks$`),
-		regexp.MustCompile(`^set_callback$`),
-		regexp.MustCompile(`^skip_callback$`),
-		// ActionDispatch middleware-stack DSL.
-		regexp.MustCompile(`^add_middleware$`),
-		regexp.MustCompile(`^delete_middleware$`),
-		regexp.MustCompile(`^insert_before$`),
-		regexp.MustCompile(`^insert_after$`),
-	}
-
-	jvmDynamicPatterns = []*regexp.Regexp{
-		// Bare-identifier forms: Java/Kotlin extractors emit only the
-		// leaf callee identifier (e.g. `m.invoke(target, args)` →
-		// ToID="invoke"). Real reflection in the wild stores the
-		// reflective handle in a local var (`Method m = clazz.getMethod(name); m.invoke(...)`),
-		// so the receiver-typed pattern `Method.invoke(` never sees the
-		// type and the call lands in BugExtractor instead of Dynamic
-		// (issue #72). The bare-name anchors below promote those leaf
-		// callees into the Dynamic disposition.
-		//
-		// Collisions with non-reflective user methods (`cli.invoke(...)`,
-		// `factory.newInstance()`) are accepted: Dynamic is the
-		// appropriate "we know it's dispatch we can't statically resolve"
-		// bucket — these stubs are equally unresolvable either way, and
-		// classifying them as Dynamic keeps them out of bug-extractor.
-		// The per-language gate (Java/Kotlin/Scala/JVM only) keeps these
-		// names from polluting other languages.
-		regexp.MustCompile(`^forName$`),                 // Class.forName
-		regexp.MustCompile(`^invoke$`),                  // Method.invoke / Constructor handle
-		regexp.MustCompile(`^newInstance$`),             // Class.newInstance / Constructor.newInstance
-		regexp.MustCompile(`^getClass$`),                // Object.getClass — reflection entry point
-		regexp.MustCompile(`^getMethod$`),               // Class.getMethod
-		regexp.MustCompile(`^getMethods$`),              // Class.getMethods
-		regexp.MustCompile(`^getDeclaredMethod$`),       // Class.getDeclaredMethod
-		regexp.MustCompile(`^getDeclaredMethods$`),      // Class.getDeclaredMethods
-		regexp.MustCompile(`^getField$`),                // Class.getField
-		regexp.MustCompile(`^getFields$`),               // Class.getFields
-		regexp.MustCompile(`^getDeclaredField$`),        // Class.getDeclaredField
-		regexp.MustCompile(`^getDeclaredFields$`),       // Class.getDeclaredFields
-		regexp.MustCompile(`^getConstructor$`),          // Class.getConstructor
-		regexp.MustCompile(`^getConstructors$`),         // Class.getConstructors
-		regexp.MustCompile(`^getDeclaredConstructor$`),  // Class.getDeclaredConstructor
-		regexp.MustCompile(`^getDeclaredConstructors$`), // Class.getDeclaredConstructors
-		// JVM reflection invoke is `Method.invoke(...)` or
-		// `Constructor.invoke(...)`. Anchored to those receivers when
-		// the full call expression is present (extractor pre-strip
-		// stubs) so a user-defined `cli.invoke(...)` / `cmd.invoke(...)`
-		// does NOT match.
-		regexp.MustCompile(`\b(?:Method|Constructor)\.invoke\(`),
-		regexp.MustCompile(`^Class\.forName\(`), // Class.forName("...")
-		// Anchored to the reflective `Class.forName(...).newInstance()` /
-		// `<Type>.class.newInstance()` shape so a plain factory method
-		// named `newInstance()` on a domain class does NOT match.
-		regexp.MustCompile(`Class\.forName\([^)]*\)\.newInstance\(`),
-		regexp.MustCompile(`\.class\.newInstance\(`),
-		regexp.MustCompile(`^ServiceLoader\.load\(`), // ServiceLoader.load(...)
-		regexp.MustCompile(`^System\.getenv\(`),      // env-driven (JVM)
-		// Issue #44 — Spring MVC / WebFlux ResponseEntity fluent builder
-		// methods. Spring's ResponseEntity uses a static factory + builder
-		// pattern:
-		//
-		//   ResponseEntity.notFound().build()
-		//   ResponseEntity.ok(body)
-		//   ResponseEntity.status(HttpStatus.CREATED).body(x)
-		//   ResponseEntity.noContent().build()
-		//
-		// The Kotlin/Java extractor emits the trailing method leaf as a
-		// bare CALLS stub (e.g. `notFound`, `ok`, `build`, `body`).
-		// Without full type-inference the resolver cannot tell whether
-		// the caller's receiver is a ResponseEntity builder or a same-
-		// named in-tree method. These names are also generic enough
-		// (`ok`, `build`, `body`) that user-defined methods can share
-		// them — both cases are statically unresolvable and Dynamic is the
-		// correct bucket. The JVM-language gate keeps these from
-		// polluting non-JVM graphs (#94 safer-bias rule).
-		regexp.MustCompile(`^notFound$`),            // ResponseEntity.notFound()
-		regexp.MustCompile(`^noContent$`),           // ResponseEntity.noContent()
-		regexp.MustCompile(`^badRequest$`),          // ResponseEntity.badRequest()
-		regexp.MustCompile(`^accepted$`),            // ResponseEntity.accepted()
-		regexp.MustCompile(`^created$`),             // ResponseEntity.created(uri)
-		regexp.MustCompile(`^unprocessableEntity$`), // ResponseEntity.unprocessableEntity()
-		regexp.MustCompile(`^internalServerError$`), // ResponseEntity.internalServerError()
-		// Builder terminal methods that appear as bare leaf stubs when
-		// the receiver is a ResponseEntity.BodyBuilder /
-		// ResponseEntity.HeadersBuilder (or any other fluent builder that
-		// can't be resolved by name alone).
-		regexp.MustCompile(`^build$`), // BodyBuilder.build() / HeadersBuilder.build()
-		regexp.MustCompile(`^body$`),  // BodyBuilder.body(T)
-		// ResponseEntity.ok(T) is already common in user code; but since
-		// the receiver is always the static class itself in Spring usage
-		// and the name is generic enough to be in-tree we accept the
-		// safer-bias trade-off — Dynamic for both cases.
-		regexp.MustCompile(`^ok$`), // ResponseEntity.ok(body)
-		// Issue #44 — Scala stdlib companion-object and collection method
-		// stubs. The Scala extractor emits qualified CALLS edges of the form
-		// "Future.successful", "List.map", "Map.get", etc. when the receiver
-		// is a stdlib type and full type inference is unavailable.
-		//
-		// These are statically unresolvable because:
-		//  1. Future / Try / Option / List / Map / Seq / Set / Vector are
-		//     external types (scala.concurrent / scala.util / scala.collection).
-		//  2. The extractor already emits the PascalCase-qualified form, so
-		//     the resolver correctly identifies the receiver but can't bind
-		//     to an entity (none is emitted for stdlib).
-		//
-		// Pattern: "^<ScalaStdlibType>\.<method>$" — qualified leaf form.
-		// JVM-language gate keeps these from polluting Python / Go / JS / Ruby.
-		//
-		// scala.concurrent.Future companion methods.
-		regexp.MustCompile(`^Future\.successful$`), // Future.successful(v)
-		regexp.MustCompile(`^Future\.failed$`),     // Future.failed(ex)
-		regexp.MustCompile(`^Future\.apply$`),      // Future { ... }
-		regexp.MustCompile(`^Future\.sequence$`),   // Future.sequence(list)
-		regexp.MustCompile(`^Future\.traverse$`),   // Future.traverse(coll)(f)
-		regexp.MustCompile(`^Future\.unit$`),       // Future.unit
-		// scala.util.Try / Success / Failure companion methods.
-		regexp.MustCompile(`^Try\.apply$`),     // Try { ... }
-		regexp.MustCompile(`^Success\.apply$`), // Success(v)
-		regexp.MustCompile(`^Failure\.apply$`), // Failure(ex)
-		// scala.collection.immutable.List companion + instance methods.
-		regexp.MustCompile(`^List\.apply$`),        // List(a, b, c)
-		regexp.MustCompile(`^List\.empty$`),        // List.empty
-		regexp.MustCompile(`^List\.from$`),         // List.from(iter)
-		regexp.MustCompile(`^List\.map$`),          // list.map(f) — qualified via receiver
-		regexp.MustCompile(`^List\.flatMap$`),      // list.flatMap(f)
-		regexp.MustCompile(`^List\.filter$`),       // list.filter(p)
-		regexp.MustCompile(`^List\.filterNot$`),    // list.filterNot(p)
-		regexp.MustCompile(`^List\.find$`),         // list.find(p)
-		regexp.MustCompile(`^List\.foldLeft$`),     // list.foldLeft(z)(f)
-		regexp.MustCompile(`^List\.foldRight$`),    // list.foldRight(z)(f)
-		regexp.MustCompile(`^List\.foreach$`),      // list.foreach(f)
-		regexp.MustCompile(`^List\.collect$`),      // list.collect(pf)
-		regexp.MustCompile(`^List\.exists$`),       // list.exists(p)
-		regexp.MustCompile(`^List\.forall$`),       // list.forall(p)
-		regexp.MustCompile(`^List\.head$`),         // list.head
-		regexp.MustCompile(`^List\.tail$`),         // list.tail
-		regexp.MustCompile(`^List\.size$`),         // list.size
-		regexp.MustCompile(`^List\.length$`),       // list.length
-		regexp.MustCompile(`^List\.isEmpty$`),      // list.isEmpty
-		regexp.MustCompile(`^List\.nonEmpty$`),     // list.nonEmpty
-		regexp.MustCompile(`^List\.toList$`),       // list.toList
-		regexp.MustCompile(`^List\.toSet$`),        // list.toSet
-		regexp.MustCompile(`^List\.toSeq$`),        // list.toSeq
-		regexp.MustCompile(`^List\.toVector$`),     // list.toVector
-		regexp.MustCompile(`^List\.sorted$`),       // list.sorted
-		regexp.MustCompile(`^List\.sortBy$`),       // list.sortBy(f)
-		regexp.MustCompile(`^List\.groupBy$`),      // list.groupBy(f)
-		regexp.MustCompile(`^List\.take$`),         // list.take(n)
-		regexp.MustCompile(`^List\.drop$`),         // list.drop(n)
-		regexp.MustCompile(`^List\.zip$`),          // list.zip(other)
-		regexp.MustCompile(`^List\.zipWithIndex$`), // list.zipWithIndex
-		regexp.MustCompile(`^List\.distinct$`),     // list.distinct
-		regexp.MustCompile(`^List\.flatten$`),      // list.flatten
-		regexp.MustCompile(`^List\.mkString$`),     // list.mkString(sep)
-		// scala.collection.immutable.Map companion + instance methods.
-		regexp.MustCompile(`^Map\.apply$`),      // Map(k -> v)
-		regexp.MustCompile(`^Map\.empty$`),      // Map.empty
-		regexp.MustCompile(`^Map\.from$`),       // Map.from(iter)
-		regexp.MustCompile(`^Map\.get$`),        // map.get(k)
-		regexp.MustCompile(`^Map\.contains$`),   // map.contains(k)
-		regexp.MustCompile(`^Map\.keys$`),       // map.keys
-		regexp.MustCompile(`^Map\.values$`),     // map.values
-		regexp.MustCompile(`^Map\.updated$`),    // map.updated(k, v)
-		regexp.MustCompile(`^Map\.removed$`),    // map.removed(k)
-		regexp.MustCompile(`^Map\.map$`),        // map.map(f)
-		regexp.MustCompile(`^Map\.filter$`),     // map.filter(p)
-		regexp.MustCompile(`^Map\.filterKeys$`), // map.filterKeys(p)
-		regexp.MustCompile(`^Map\.mapValues$`),  // map.mapValues(f)
-		regexp.MustCompile(`^Map\.toList$`),     // map.toList
-		regexp.MustCompile(`^Map\.toSeq$`),      // map.toSeq
-		regexp.MustCompile(`^Map\.size$`),       // map.size
-		regexp.MustCompile(`^Map\.isEmpty$`),    // map.isEmpty
-		// scala.collection.immutable.Seq / Vector / Set mirrors.
-		regexp.MustCompile(`^Seq\.apply$`), // Seq(...)
-		regexp.MustCompile(`^Seq\.empty$`), // Seq.empty
-		regexp.MustCompile(`^Seq\.map$`),
-		regexp.MustCompile(`^Seq\.flatMap$`),
-		regexp.MustCompile(`^Seq\.filter$`),
-		regexp.MustCompile(`^Seq\.filterNot$`),
-		regexp.MustCompile(`^Seq\.find$`),
-		regexp.MustCompile(`^Seq\.foreach$`),
-		regexp.MustCompile(`^Vector\.apply$`), // Vector(...)
-		regexp.MustCompile(`^Vector\.empty$`), // Vector.empty
-		regexp.MustCompile(`^Vector\.from$`),  // Vector.from(iter)
-		regexp.MustCompile(`^Set\.apply$`),    // Set(...)
-		regexp.MustCompile(`^Set\.empty$`),    // Set.empty
-		regexp.MustCompile(`^Set\.from$`),     // Set.from(iter)
-		// scala.Option instance and companion methods.
-		regexp.MustCompile(`^Option\.apply$`), // Option(v)
-		regexp.MustCompile(`^Option\.empty$`), // Option.empty / None
-		regexp.MustCompile(`^Some\.apply$`),   // Some(v)
-		regexp.MustCompile(`^None\.get$`),     // None.get (rare but emitted)
-	}
-
-	// HCL / Terraform dynamic-pattern catalog (issue #44). Terraform
-	// interpolations bind at apply time through provider / module / variable
-	// indirection that no static resolver can fully follow. The HCL
-	// extractor already emits same-file structural-refs for `var.*`,
-	// `local.*`, `module.*`, `data.*.*`, `output.*`, and resource refs;
-	// what lands here are the residual cross-file refs inside a
-	// multi-file module (root module's `variables.tf` referenced from
-	// `main.tf`, etc.), provider/meta-arg dispatches, and Terraform
-	// built-in function leaves the extractor surfaces as bare callees.
-	// Per-language gate (lang == "hcl" or "terraform") keeps these
-	// patterns from poisoning resolution in other ecosystems where
-	// `local`, `module`, `data`, `var`, `count`, `for_each`, `merge`,
-	// `lookup`, etc. are common identifiers.
-	hclDynamicPatterns = []*regexp.Regexp{
-		// Terraform built-in reference prefixes. Same-file structural-refs
-		// resolve via byLocation; cross-file leftovers are framework
-		// dispatch the resolver cannot statically bind.
-		regexp.MustCompile(`^var\.[A-Za-z_]`),      // var.<name>
-		regexp.MustCompile(`^local\.[A-Za-z_]`),    // local.<name>
-		regexp.MustCompile(`^module\.[A-Za-z_]`),   // module.<name>(.attr)
-		regexp.MustCompile(`^data\.[A-Za-z_]`),     // data.<type>.<name>(.attr)
-		regexp.MustCompile(`^output\.[A-Za-z_]`),   // output.<name>
-		regexp.MustCompile(`^provider\.[A-Za-z_]`), // provider.<name>
-		regexp.MustCompile(`^path\.(module|root|cwd)`),
-		regexp.MustCompile(`^terraform\.(workspace|env)`),
-		regexp.MustCompile(`^self\.`), // self.<attr> inside provisioner blocks
-		regexp.MustCompile(`^each\.(key|value)`),
-		regexp.MustCompile(`^count\.index`),
-		// `dynamic "<label>" { ... }` blocks introduce an iteration symbol
-		// equal to the label, so `dynamic "statement" { ... }` produces
-		// `statement.value` / `statement.key` references. The label names
-		// are arbitrary user identifiers; the suffix is what marks the
-		// reference as Terraform iteration dispatch.
-		regexp.MustCompile(`^[a-z_][a-z0-9_]*\.(value|key)$`),
-		// `for x in <expr> : ...` and `[for x in <expr> : x.<attr>]`
-		// introduce a single-letter (or short) iteration variable used as
-		// `<iter>.<attr>`. Anchored to a single lowercase letter followed
-		// by a dot so user identifiers like `aws_instance.x` are not
-		// swept up by accident.
-		regexp.MustCompile(`^[a-z]\.[a-z_]`),
-		// Terraform meta-arguments arriving as bare leaves.
-		regexp.MustCompile(`^count$`),
-		regexp.MustCompile(`^for_each$`),
-		regexp.MustCompile(`^depends_on$`),
-		regexp.MustCompile(`^lifecycle$`),
-		regexp.MustCompile(`^dynamic$`),
-		regexp.MustCompile(`^provisioner$`),
-		regexp.MustCompile(`^connection$`),
-		// Terraform built-in function names (full catalog).
-		// Numeric.
-		regexp.MustCompile(`^abs$`),
-		regexp.MustCompile(`^ceil$`),
-		regexp.MustCompile(`^floor$`),
-		regexp.MustCompile(`^log$`),
-		regexp.MustCompile(`^max$`),
-		regexp.MustCompile(`^min$`),
-		regexp.MustCompile(`^parseint$`),
-		regexp.MustCompile(`^pow$`),
-		regexp.MustCompile(`^signum$`),
-		// String.
-		regexp.MustCompile(`^chomp$`),
-		regexp.MustCompile(`^format$`),
-		regexp.MustCompile(`^formatlist$`),
-		regexp.MustCompile(`^indent$`),
-		regexp.MustCompile(`^join$`),
-		regexp.MustCompile(`^lower$`),
-		regexp.MustCompile(`^regex$`),
-		regexp.MustCompile(`^regexall$`),
-		regexp.MustCompile(`^replace$`),
-		regexp.MustCompile(`^split$`),
-		regexp.MustCompile(`^strrev$`),
-		regexp.MustCompile(`^substr$`),
-		regexp.MustCompile(`^title$`),
-		regexp.MustCompile(`^trim$`),
-		regexp.MustCompile(`^trimprefix$`),
-		regexp.MustCompile(`^trimsuffix$`),
-		regexp.MustCompile(`^trimspace$`),
-		regexp.MustCompile(`^upper$`),
-		regexp.MustCompile(`^startswith$`),
-		regexp.MustCompile(`^endswith$`),
-		// Collection.
-		regexp.MustCompile(`^alltrue$`),
-		regexp.MustCompile(`^anytrue$`),
-		regexp.MustCompile(`^chunklist$`),
-		regexp.MustCompile(`^coalesce$`),
-		regexp.MustCompile(`^coalescelist$`),
-		regexp.MustCompile(`^compact$`),
-		regexp.MustCompile(`^concat$`),
-		regexp.MustCompile(`^contains$`),
-		regexp.MustCompile(`^distinct$`),
-		regexp.MustCompile(`^element$`),
-		regexp.MustCompile(`^flatten$`),
-		regexp.MustCompile(`^index$`),
-		regexp.MustCompile(`^keys$`),
-		regexp.MustCompile(`^length$`),
-		regexp.MustCompile(`^list$`),
-		regexp.MustCompile(`^lookup$`),
-		regexp.MustCompile(`^map$`),
-		regexp.MustCompile(`^matchkeys$`),
-		regexp.MustCompile(`^merge$`),
-		regexp.MustCompile(`^one$`),
-		regexp.MustCompile(`^range$`),
-		regexp.MustCompile(`^reverse$`),
-		regexp.MustCompile(`^setintersection$`),
-		regexp.MustCompile(`^setproduct$`),
-		regexp.MustCompile(`^setsubtract$`),
-		regexp.MustCompile(`^setunion$`),
-		regexp.MustCompile(`^slice$`),
-		regexp.MustCompile(`^sort$`),
-		regexp.MustCompile(`^sum$`),
-		regexp.MustCompile(`^transpose$`),
-		regexp.MustCompile(`^try$`),
-		regexp.MustCompile(`^values$`),
-		regexp.MustCompile(`^zipmap$`),
-		regexp.MustCompile(`^nonsensitive$`),
-		regexp.MustCompile(`^sensitive$`),
-		// Encoding.
-		regexp.MustCompile(`^base64decode$`),
-		regexp.MustCompile(`^base64encode$`),
-		regexp.MustCompile(`^base64gzip$`),
-		regexp.MustCompile(`^csvdecode$`),
-		regexp.MustCompile(`^jsondecode$`),
-		regexp.MustCompile(`^jsonencode$`),
-		regexp.MustCompile(`^textdecodebase64$`),
-		regexp.MustCompile(`^textencodebase64$`),
-		regexp.MustCompile(`^urlencode$`),
-		regexp.MustCompile(`^yamldecode$`),
-		regexp.MustCompile(`^yamlencode$`),
-		// Filesystem / template.
-		regexp.MustCompile(`^abspath$`),
-		regexp.MustCompile(`^basename$`),
-		regexp.MustCompile(`^dirname$`),
-		regexp.MustCompile(`^pathexpand$`),
-		regexp.MustCompile(`^file$`),
-		regexp.MustCompile(`^fileexists$`),
-		regexp.MustCompile(`^fileset$`),
-		regexp.MustCompile(`^filebase64$`),
-		regexp.MustCompile(`^templatefile$`),
-		regexp.MustCompile(`^templatestring$`),
-		// Date/time.
-		regexp.MustCompile(`^formatdate$`),
-		regexp.MustCompile(`^plantimestamp$`),
-		regexp.MustCompile(`^timeadd$`),
-		regexp.MustCompile(`^timecmp$`),
-		regexp.MustCompile(`^timestamp$`),
-		// Hash/crypto.
-		regexp.MustCompile(`^base64sha256$`),
-		regexp.MustCompile(`^base64sha512$`),
-		regexp.MustCompile(`^bcrypt$`),
-		regexp.MustCompile(`^filebase64sha256$`),
-		regexp.MustCompile(`^filebase64sha512$`),
-		regexp.MustCompile(`^filemd5$`),
-		regexp.MustCompile(`^filesha1$`),
-		regexp.MustCompile(`^filesha256$`),
-		regexp.MustCompile(`^filesha512$`),
-		regexp.MustCompile(`^md5$`),
-		regexp.MustCompile(`^rsadecrypt$`),
-		regexp.MustCompile(`^sha1$`),
-		regexp.MustCompile(`^sha256$`),
-		regexp.MustCompile(`^sha512$`),
-		regexp.MustCompile(`^uuid$`),
-		regexp.MustCompile(`^uuidv5$`),
-		// IP/network.
-		regexp.MustCompile(`^cidrhost$`),
-		regexp.MustCompile(`^cidrnetmask$`),
-		regexp.MustCompile(`^cidrsubnet$`),
-		regexp.MustCompile(`^cidrsubnets$`),
-		// Type conversion.
-		regexp.MustCompile(`^can$`),
-		regexp.MustCompile(`^tobool$`),
-		regexp.MustCompile(`^tolist$`),
-		regexp.MustCompile(`^tomap$`),
-		regexp.MustCompile(`^tonumber$`),
-		regexp.MustCompile(`^toset$`),
-		regexp.MustCompile(`^tostring$`),
-		regexp.MustCompile(`^type$`),
-		// Variable-shaped leaves that arrive bare from the extractor when
-		// an interpolation is just `var.foo` (no further .attr) — already
-		// covered by `^var\.` etc above; the bare `var`, `local`, `module`
-		// leaves (rare, e.g. via dynamic blocks) round it out.
-		regexp.MustCompile(`^var$`),
-		regexp.MustCompile(`^local$`),
-		regexp.MustCompile(`^module$`),
-		regexp.MustCompile(`^data$`),
-		regexp.MustCompile(`^output$`),
-		// Provider-prefixed resource / data refs that miss the same-file
-		// structural-ref bind (declared in a sibling file of the same
-		// module). The major Terraform provider prefixes — any new
-		// provider follows the same `<prefix>_<resource_type>.<name>`
-		// shape, so anchoring on the prefix keeps the catalog stable
-		// without enumerating every provider.
-		regexp.MustCompile(`^aws_[a-z0-9_]+\.`),
-		regexp.MustCompile(`^azurerm_[a-z0-9_]+\.`),
-		regexp.MustCompile(`^azuread_[a-z0-9_]+\.`),
-		regexp.MustCompile(`^google_[a-z0-9_]+\.`),
-		regexp.MustCompile(`^kubernetes_[a-z0-9_]+\.`),
-		regexp.MustCompile(`^helm_[a-z0-9_]+\.`),
-		regexp.MustCompile(`^oci_[a-z0-9_]+\.`),
-		regexp.MustCompile(`^null_[a-z0-9_]+\.`),
-		regexp.MustCompile(`^random_[a-z0-9_]+\.`),
-		regexp.MustCompile(`^tls_[a-z0-9_]+\.`),
-		regexp.MustCompile(`^template_[a-z0-9_]+\.`),
-		regexp.MustCompile(`^archive_[a-z0-9_]+\.`),
-		regexp.MustCompile(`^external_[a-z0-9_]+\.`),
-		regexp.MustCompile(`^http_[a-z0-9_]+\.`),
-		regexp.MustCompile(`^vault_[a-z0-9_]+\.`),
-		regexp.MustCompile(`^datadog_[a-z0-9_]+\.`),
-		regexp.MustCompile(`^cloudflare_[a-z0-9_]+\.`),
-		regexp.MustCompile(`^github_[a-z0-9_]+\.`),
-		regexp.MustCompile(`^gitlab_[a-z0-9_]+\.`),
-		regexp.MustCompile(`^digitalocean_[a-z0-9_]+\.`),
-		regexp.MustCompile(`^linode_[a-z0-9_]+\.`),
-		regexp.MustCompile(`^alicloud_[a-z0-9_]+\.`),
-		regexp.MustCompile(`^tencentcloud_[a-z0-9_]+\.`),
-		regexp.MustCompile(`^hcp_[a-z0-9_]+\.`),
-		regexp.MustCompile(`^consul_[a-z0-9_]+\.`),
-		regexp.MustCompile(`^nomad_[a-z0-9_]+\.`),
-		regexp.MustCompile(`^docker_[a-z0-9_]+\.`),
-		// Bare provider names — emitted as IMPORTS ToID by the HCL
-		// extractor for `provider "<name>"` blocks. Same rationale as the
-		// resource-prefix patterns above: the provider plugin lives outside
-		// the static graph and the dynamic bucket is the right disposition.
-		regexp.MustCompile(`^aws$`),
-		regexp.MustCompile(`^azurerm$`),
-		regexp.MustCompile(`^azuread$`),
-		regexp.MustCompile(`^google$`),
-		regexp.MustCompile(`^kubernetes$`),
-		regexp.MustCompile(`^helm$`),
-		regexp.MustCompile(`^oci$`),
-		regexp.MustCompile(`^null$`),
-		regexp.MustCompile(`^random$`),
-		regexp.MustCompile(`^tls$`),
-		regexp.MustCompile(`^template$`),
-		regexp.MustCompile(`^archive$`),
-		regexp.MustCompile(`^external$`),
-		regexp.MustCompile(`^http$`),
-		regexp.MustCompile(`^vault$`),
-		regexp.MustCompile(`^datadog$`),
-		regexp.MustCompile(`^cloudflare$`),
-		regexp.MustCompile(`^github$`),
-		regexp.MustCompile(`^gitlab$`),
-		regexp.MustCompile(`^digitalocean$`),
-		regexp.MustCompile(`^linode$`),
-		regexp.MustCompile(`^alicloud$`),
-		regexp.MustCompile(`^tencentcloud$`),
-		regexp.MustCompile(`^hcp$`),
-		regexp.MustCompile(`^consul$`),
-		regexp.MustCompile(`^nomad$`),
-		regexp.MustCompile(`^docker$`),
-		// Relative-path module sources (`source = "../../"`,
-		// `source = "./modules/foo"`). The HCL extractor emits these as
-		// raw IMPORTS ToIDs for module blocks. Local module paths could
-		// be resolved to a sibling-directory entity in principle but the
-		// pattern of `..`/`./` prefixes is unambiguously a path; tagging
-		// Dynamic is consistent with how Python's relative-import paths
-		// (`^\.+`) land in pythonDynamicPatterns.
-		regexp.MustCompile(`^\.\.?/`),
-		regexp.MustCompile(`^\.\.?$`),
-		// Terraform registry module sources
-		// (`registry.terraform.io/hashicorp/aws/version`). External
-		// package not in the graph. The 3-segment registry short form
-		// (`hashicorp/aws/version`) is deliberately NOT pattern-matched
-		// here — it collides with markdown IMPORTS shaped the same way
-		// (`docs/modules/vpc-endpoints`); the long form (with the
-		// `registry.terraform.io/` host) is unambiguous.
-		regexp.MustCompile(`^registry\.terraform\.io/`),
-		regexp.MustCompile(`^git::`),
-		regexp.MustCompile(`^github\.com/`),
-		regexp.MustCompile(`^bitbucket\.org/`),
-	}
-
-	// Rust dynamic-pattern catalog (issue #44 slice-7).
-	//
-	// The Rust extractor emits two shapes that can never be resolved
-	// internally and land in bug-extractor without these patterns:
-	//
-	//  1. Bare channel-constructor names.  `mpsc::channel::<String>(8)` is
-	//     a `generic_function > scoped_identifier` call site; the extractor
-	//     strips the `mpsc::` module path and emits the bare leaf `channel`.
-	//     The same pattern covers `oneshot::channel`, `broadcast::channel`,
-	//     and `watch::channel` — all tokio / std concurrency primitives with
-	//     no in-tree entity.
-	//
-	//  2. Generic-receiver method calls.  When a local variable `rx` is
-	//     declared as `Receiver<String>` (or any `Type<T>`) the extractor
-	//     emits the stub as `Receiver<String>.recv` — the fully-qualified
-	//     receiver shape including the generic argument.  No in-tree entity
-	//     can ever satisfy that exact to_id because the resolver's bare-name
-	//     lookup strips generics before matching.  These are stdlib / tokio
-	//     concrete-type method calls (recv, send, close, poll, next, …) that
-	//     are intrinsically unresolvable statically — Dynamic is correct.
-	//
-	// Both patterns are gated to lang=="rust" so they cannot fire on
-	// PascalCase method names in other languages (safer-bias rule #94).
-	rustDynamicPatterns = []*regexp.Regexp{
-		// Bare channel-constructor stubs: mpsc::channel, oneshot::channel,
-		// broadcast::channel, watch::channel, etc. The extractor preserves
-		// only the leaf name after the last `::`.
-		regexp.MustCompile(`^channel$`),
-		// Generic-receiver method stubs: `Type<T>.method` where the type
-		// carries at least one generic argument.  The extractor qualifies the
-		// call with the full concrete receiver type including `<...>` which
-		// the resolver cannot match against any entity.  The inner generic
-		// argument may itself be generic (`Arc<Mutex<State>>`), so the match
-		// uses a greedy `.+` inside `<…>` rather than `[^>]+`.
-		regexp.MustCompile(`^[A-Z][A-Za-z0-9_]*<.+>\.[a-z_][A-Za-z0-9_]*$`),
-	}
-
-// Lua dynamic-pattern catalog (issue #44 — Lua resolver slice).
-	//
-	// The Lua extractor's luaCallTarget helper returns the LAST identifier
-	// before the argument list for any function_call node.  For a dotted
-	// stdlib call like `table.insert(t, v)` the receiver ("table") is
-	// discarded and only the leaf callee ("insert") reaches the resolver.
-	// The same stripping happens for `string.format`, `math.floor`,
-	// `os.getenv`, `coroutine.resume`, etc.
-	//
-	// Two categories are handled:
-	//
-	//  1. Lua-unique global built-ins — identifiers that are part of the
-	//     Lua standard environment and are virtually never user-defined
-	//     function names in any other language:
-	//       ipairs, pairs, pcall, xpcall, rawget, rawset, rawequal, rawlen,
-	//       setmetatable, getmetatable, tostring, tonumber, unpack, select,
-	//       coroutine.wrap → bare "wrap" (Lua-gate makes this safe enough).
-	//     These names land as BugExtractor on every Lua corpus because the
-	//     graph never contains an entity whose Name is "ipairs". With the
-	//     Lua language gate the risk of false-positive against a user-defined
-	//     `ipairs`/`pairs`/`pcall` is negligible.
-	//
-	//  2. Lua stdlib module leaf names — table.*, string.*, math.*, os.*,
-	//     coroutine.* leaf identifiers that are too common as user-defined
-	//     method names to be safe cross-language, but are safe under the
-	//     Lua gate because real Lua code uses them exclusively as stdlib
-	//     calls.  Covered: insert, remove, sort, concat (table.*);
-	//     format, sub, gmatch, gsub, find, byte, char, rep, len
-	//     (string.*); floor, ceil, sqrt, abs, max, min, random, huge, pi,
-	//     fmod, modf, exp, log (math.*); getenv, time, clock, exit, date,
-	//     execute (os.*); resume, yield, status, create, wrap (coroutine.*).
-	//
-	// All patterns are gated to lang=="lua" so they cannot fire on stubs
-	// from Go, Python, Ruby, Java, TypeScript, or any other language
-	// (safer-bias rule #94).  Cross-language collisions that motivated this
-	// gate: `format` (Python str.format), `insert` (SQL), `remove` (JS
-	// Array.prototype.remove), `exit` (C/Python), `type` (TypeScript),
-	// `create` (factory pattern in any language).
-	luaDynamicPatterns = []*regexp.Regexp{
-		// ── Tier 1: Lua-unique global identifiers ──────────────────────
-		// These names exist in no other language's standard global scope
-		// with this exact spelling; the Lua gate is a belt-and-suspenders
-		// measure rather than a primary safety control.
-		regexp.MustCompile(`^ipairs$`),
-		regexp.MustCompile(`^pairs$`),
-		regexp.MustCompile(`^pcall$`),
-		regexp.MustCompile(`^xpcall$`),
-		regexp.MustCompile(`^rawget$`),
-		regexp.MustCompile(`^rawset$`),
-		regexp.MustCompile(`^rawequal$`),
-		regexp.MustCompile(`^rawlen$`),
-		regexp.MustCompile(`^setmetatable$`),
-		regexp.MustCompile(`^getmetatable$`),
-		regexp.MustCompile(`^tostring$`),
-		regexp.MustCompile(`^tonumber$`),
-		regexp.MustCompile(`^unpack$`),
-		regexp.MustCompile(`^select$`),
-		regexp.MustCompile(`^next$`),
-		regexp.MustCompile(`^collectgarbage$`),
-		regexp.MustCompile(`^dofile$`),
-		regexp.MustCompile(`^loadfile$`),
-		regexp.MustCompile(`^loadstring$`),
-		regexp.MustCompile(`^loadchunk$`),
-
-		// ── Tier 2: table.* stdlib leaf names ─────────────────────────
-		// Gated to lua; too generic otherwise (`insert` in SQL/Java,
-		// `remove` in JS/C++, `sort` in Python/Go, `concat` in JS).
-		regexp.MustCompile(`^insert$`),
-		regexp.MustCompile(`^remove$`),
-		regexp.MustCompile(`^sort$`),
-		regexp.MustCompile(`^concat$`),
-		regexp.MustCompile(`^move$`),
-
-		// ── Tier 2: string.* stdlib leaf names ────────────────────────
-		// `format`/`sub`/`find`/`byte`/`char`/`rep`/`len`/`gmatch`/`gsub`
-		// are Lua string lib methods the extractor strips to bare leaves.
-		regexp.MustCompile(`^gmatch$`),
-		regexp.MustCompile(`^gsub$`),
-		regexp.MustCompile(`^byte$`),
-		regexp.MustCompile(`^char$`),
-		regexp.MustCompile(`^rep$`),
-		regexp.MustCompile(`^dump$`), // string.dump
-
-		// ── Tier 2: math.* stdlib leaf names ──────────────────────────
-		regexp.MustCompile(`^floor$`),
-		regexp.MustCompile(`^ceil$`),
-		regexp.MustCompile(`^sqrt$`),
-		regexp.MustCompile(`^fabs$`),
-		regexp.MustCompile(`^fmod$`),
-		regexp.MustCompile(`^modf$`),
-		regexp.MustCompile(`^frexp$`),
-		regexp.MustCompile(`^ldexp$`),
-		regexp.MustCompile(`^random$`),
-		regexp.MustCompile(`^randomseed$`),
-		regexp.MustCompile(`^huge$`),
-		regexp.MustCompile(`^tointeger$`), // math.tointeger (Lua 5.3+)
-
-		// ── Tier 2: os.* stdlib leaf names ────────────────────────────
-		regexp.MustCompile(`^tmpname$`),  // os.tmpname
-		regexp.MustCompile(`^difftime$`), // os.difftime
-
-		// ── Tier 2: coroutine.* stdlib leaf names ─────────────────────
-		// `resume`, `yield`, `wrap`, `status`, `running`, `isyieldable`
-		// are coroutine module methods.  `create` is also used here but is
-		// too ambiguous even with a Lua gate (factory pattern), so it is
-		// deliberately excluded.
-		regexp.MustCompile(`^resume$`),
-		regexp.MustCompile(`^yield$`),
-		regexp.MustCompile(`^isyieldable$`), // Lua 5.3+
-		regexp.MustCompile(`^running$`),     // coroutine.running
-	}
-
-		// Swift dynamic-pattern catalog (issue #44). The Swift extractor
-	// (internal/extractors/swift/swift.go) emits CALLS edges whose ToID
-	// is the bare trailing method identifier extracted from navigation
-	// expressions — e.g. `publisher.sink(...)` → ToID="sink",
-	// `url.appendingPathComponent("users")` → ToID="appendingPathComponent".
-	// When the receiver is an external framework type (Combine Publisher,
-	// Foundation URL/URLSession, SwiftUI View) no in-tree entity exists
-	// for the callee, so the resolver cannot bind it. These stubs are
-	// statically unresolvable because:
-	//   1. The Combine / Foundation / SwiftUI types are imported from the
-	//      Apple SDK — they are never indexed as in-tree entities.
-	//   2. The extractor emits only the leaf name; without the receiver
-	//      type the resolver cannot distinguish `publisher.sink` from a
-	//      user-defined method also named `sink` on a domain type.
-	//
-	// The safer-bias rule (#94) is preserved by the per-language gate
-	// (lang=="swift"): identical names common in other ecosystems
-	// (`store`, `sink`, `map`, `filter`, `receive`) do NOT fire here for
-	// Go, Python, Ruby, etc.
-	//
-	// Combine publisher operators (the most common unresolved category).
-	// Apple Combine framework — AnyPublisher, Publisher, Subscriber, etc.
-	// Each operator appears as a bare leaf name on a publisher chain.
-	swiftDynamicPatterns = []*regexp.Regexp{
-		// Combine subscription lifecycle.
-		regexp.MustCompile(`^sink$`),               // publisher.sink(receiveValue:)
-		regexp.MustCompile(`^store$`),              // cancellable.store(in:)
-		regexp.MustCompile(`^cancel$`),             // AnyCancellable.cancel()
-		regexp.MustCompile(`^eraseToAnyPublisher$`), // publisher.eraseToAnyPublisher()
-		// Combine upstream operators.
-		regexp.MustCompile(`^receive$`),            // publisher.receive(on:)
-		regexp.MustCompile(`^subscribe$`),          // publisher.subscribe(_:)
-		regexp.MustCompile(`^connect$`),            // ConnectablePublisher.connect()
-		regexp.MustCompile(`^autoconnect$`),        // ConnectablePublisher.autoconnect()
-		regexp.MustCompile(`^share$`),              // publisher.share()
-		regexp.MustCompile(`^multicast$`),          // publisher.multicast(_:)
-		// Combine transformation operators.
-		regexp.MustCompile(`^mapError$`),           // publisher.mapError(_:)
-		regexp.MustCompile(`^flatMap$`),            // publisher.flatMap(_:)
-		regexp.MustCompile(`^compactMap$`),         // publisher.compactMap(_:)
-		regexp.MustCompile(`^tryMap$`),             // publisher.tryMap(_:)
-		regexp.MustCompile(`^tryFlatMap$`),         // publisher.tryFlatMap(_:)
-		regexp.MustCompile(`^tryCompactMap$`),      // publisher.tryCompactMap(_:)
-		regexp.MustCompile(`^tryFilter$`),          // publisher.tryFilter(_:)
-		regexp.MustCompile(`^decode$`),             // publisher.decode(type:decoder:)
-		regexp.MustCompile(`^encode$`),             // publisher.encode(encoder:)
-		regexp.MustCompile(`^scan$`),               // publisher.scan(_:_:)
-		regexp.MustCompile(`^tryScan$`),            // publisher.tryScan(_:_:)
-		regexp.MustCompile(`^reduce$`),             // publisher.reduce(_:_:)
-		regexp.MustCompile(`^tryReduce$`),          // publisher.tryReduce(_:_:)
-		regexp.MustCompile(`^collect$`),            // publisher.collect() / .collect(_:)
-		regexp.MustCompile(`^replaceNil$`),         // publisher.replaceNil(with:)
-		regexp.MustCompile(`^replaceEmpty$`),       // publisher.replaceEmpty(with:)
-		regexp.MustCompile(`^replaceError$`),       // publisher.replaceError(with:)
-		regexp.MustCompile(`^setFailureType$`),     // publisher.setFailureType(to:)
-		// Combine filtering operators.
-		regexp.MustCompile(`^removeDuplicates$`),   // publisher.removeDuplicates()
-		regexp.MustCompile(`^debounce$`),           // publisher.debounce(for:scheduler:)
-		regexp.MustCompile(`^throttle$`),           // publisher.throttle(for:scheduler:latest:)
-		regexp.MustCompile(`^timeout$`),            // publisher.timeout(_:scheduler:)
-		regexp.MustCompile(`^retry$`),              // publisher.retry(_:)
-		regexp.MustCompile(`^catchError$`),         // publisher.catch(_:)
-		regexp.MustCompile(`^tryCatch$`),           // publisher.tryCatch(_:)
-		regexp.MustCompile(`^first$`),              // publisher.first() / .first(where:)
-		regexp.MustCompile(`^tryFirst$`),           // publisher.tryFirst(where:)
-		regexp.MustCompile(`^last$`),               // publisher.last() / .last(where:)
-		regexp.MustCompile(`^tryLast$`),            // publisher.tryLast(where:)
-		regexp.MustCompile(`^ignoreOutput$`),       // publisher.ignoreOutput()
-		regexp.MustCompile(`^prefix$`),             // publisher.prefix(_:)
-		regexp.MustCompile(`^dropFirst$`),          // publisher.dropFirst(_:)
-		regexp.MustCompile(`^drop$`),               // publisher.drop(untilOutputFrom:)
-		regexp.MustCompile(`^makeConnectable$`),    // publisher.makeConnectable()
-		// Combine combining operators.
-		regexp.MustCompile(`^zip$`),                // Publishers.Zip / publisher.zip(_:)
-		regexp.MustCompile(`^combineLatest$`),      // publisher.combineLatest(_:)
-		regexp.MustCompile(`^merge$`),              // publisher.merge(with:)
-		regexp.MustCompile(`^switchToLatest$`),     // publisher.switchToLatest()
-		regexp.MustCompile(`^prepend$`),            // publisher.prepend(_:)
-		regexp.MustCompile(`^append$`),             // publisher.append(_:)
-		// Combine side-effect operators.
-		regexp.MustCompile(`^handleEvents$`),       // publisher.handleEvents(...)
-		regexp.MustCompile(`^breakpoint$`),         // publisher.breakpoint()
-		regexp.MustCompile(`^breakpointOnError$`),  // publisher.breakpointOnError()
-		regexp.MustCompile(`^print$`),              // publisher.print(_:)
-		regexp.MustCompile(`^measureInterval$`),    // publisher.measureInterval(using:)
-		// Combine subject send / value methods.
-		regexp.MustCompile(`^send$`),               // subject.send(_:)
-		regexp.MustCompile(`^sendCompletion$`),     // subject.send(completion:)
-		// Combine assign operator.
-		regexp.MustCompile(`^assign$`),             // publisher.assign(to:on:)
-		// Foundation / URLSession + URL methods that appear as bare
-		// leaf names from navigation chains on external framework types.
-		// These are Apple SDK methods — no in-tree entity is indexed.
-		regexp.MustCompile(`^dataTaskPublisher$`),   // URLSession.shared.dataTaskPublisher(for:)
-		regexp.MustCompile(`^dataTask$`),            // URLSession.dataTask(with:)
-		regexp.MustCompile(`^uploadTask$`),          // URLSession.uploadTask(with:from:)
-		regexp.MustCompile(`^downloadTask$`),        // URLSession.downloadTask(with:)
-		regexp.MustCompile(`^webSocketTask$`),       // URLSession.webSocketTask(with:)
-		regexp.MustCompile(`^appendingPathComponent$`), // URL.appendingPathComponent(_:)
-		regexp.MustCompile(`^appendingPathExtension$`), // URL.appendingPathExtension(_:)
-		regexp.MustCompile(`^deletingLastPathComponent$`), // URL.deletingLastPathComponent()
-		regexp.MustCompile(`^deletingPathExtension$`), // URL.deletingPathExtension()
-		regexp.MustCompile(`^resolvingSymlinksInPath$`), // URL.resolvingSymlinksInPath()
-		regexp.MustCompile(`^standardized$`),        // URL.standardized
-		regexp.MustCompile(`^absoluteURL$`),         // URL.absoluteURL
-		regexp.MustCompile(`^setValue$`),            // URLRequest.setValue(_:forHTTPHeaderField:)
-		regexp.MustCompile(`^addValue$`),            // URLRequest.addValue(_:forHTTPHeaderField:)
-		// SwiftUI view modifier chain methods. SwiftUI views use a
-		// modifier-chaining pattern: `Text(...).font(.headline).padding()`.
-		// The Swift tree-sitter extractor extracts each modifier as a bare
-		// CALLS edge with the modifier name as ToID. SwiftUI modifiers are
-		// framework methods on `View` (an external protocol from SwiftUI
-		// module) — no in-tree entity can be indexed for them.
-		regexp.MustCompile(`^padding$`),             // .padding() / .padding(_:)
-		regexp.MustCompile(`^frame$`),               // .frame(width:height:) etc.
-		regexp.MustCompile(`^background$`),          // .background(_:)
-		regexp.MustCompile(`^foregroundColor$`),     // .foregroundColor(_:) (deprecated but common)
-		regexp.MustCompile(`^foregroundStyle$`),     // .foregroundStyle(_:)
-		regexp.MustCompile(`^font$`),                // .font(_:)
-		regexp.MustCompile(`^cornerRadius$`),        // .cornerRadius(_:)
-		regexp.MustCompile(`^clipShape$`),           // .clipShape(_:)
-		regexp.MustCompile(`^overlay$`),             // .overlay(_:)
-		regexp.MustCompile(`^shadow$`),              // .shadow(color:radius:x:y:)
-		regexp.MustCompile(`^opacity$`),             // .opacity(_:)
-		regexp.MustCompile(`^scaleEffect$`),         // .scaleEffect(_:)
-		regexp.MustCompile(`^rotationEffect$`),      // .rotationEffect(_:)
-		regexp.MustCompile(`^offset$`),              // .offset(x:y:)
-		regexp.MustCompile(`^position$`),            // .position(x:y:)
-		regexp.MustCompile(`^blur$`),                // .blur(radius:)
-		regexp.MustCompile(`^saturation$`),          // .saturation(_:)
-		regexp.MustCompile(`^brightness$`),          // .brightness(_:)
-		regexp.MustCompile(`^contrast$`),            // .contrast(_:)
-		regexp.MustCompile(`^colorInvert$`),         // .colorInvert()
-		regexp.MustCompile(`^grayscale$`),           // .grayscale(_:)
-		regexp.MustCompile(`^animation$`),           // .animation(_:)
-		regexp.MustCompile(`^transition$`),          // .transition(_:)
-		regexp.MustCompile(`^zIndex$`),              // .zIndex(_:)
-		regexp.MustCompile(`^fixedSize$`),           // .fixedSize()
-		regexp.MustCompile(`^layoutPriority$`),      // .layoutPriority(_:)
-		regexp.MustCompile(`^navigationTitle$`),     // .navigationTitle(_:)
-		regexp.MustCompile(`^navigationBarTitleDisplayMode$`), // .navigationBarTitleDisplayMode(_:)
-		regexp.MustCompile(`^navigationBarHidden$`), // .navigationBarHidden(_:)
-		regexp.MustCompile(`^navigationBarBackButtonHidden$`), // .navigationBarBackButtonHidden(_:)
-		regexp.MustCompile(`^toolbar$`),             // .toolbar { ... }
-		regexp.MustCompile(`^toolbarBackground$`),   // .toolbarBackground(_:for:)
-		regexp.MustCompile(`^toolbarColorScheme$`),  // .toolbarColorScheme(_:for:)
-		regexp.MustCompile(`^searchable$`),          // .searchable(text:)
-		regexp.MustCompile(`^onAppear$`),            // .onAppear { ... }
-		regexp.MustCompile(`^onDisappear$`),         // .onDisappear { ... }
-		regexp.MustCompile(`^onTapGesture$`),        // .onTapGesture { ... }
-		regexp.MustCompile(`^onLongPressGesture$`),  // .onLongPressGesture { ... }
-		regexp.MustCompile(`^onSubmit$`),            // .onSubmit { ... }
-		regexp.MustCompile(`^onChange$`),            // .onChange(of:) { ... }
-		regexp.MustCompile(`^onReceive$`),           // .onReceive(_:) { ... }
-		regexp.MustCompile(`^onDelete$`),            // .onDelete(perform:)
-		regexp.MustCompile(`^onMove$`),              // .onMove(perform:)
-		regexp.MustCompile(`^task$`),                // .task { ... } (async task modifier)
-		regexp.MustCompile(`^refreshable$`),         // .refreshable { ... }
-		regexp.MustCompile(`^swipeActions$`),        // .swipeActions { ... }
-		regexp.MustCompile(`^contextMenu$`),         // .contextMenu { ... }
-		regexp.MustCompile(`^sheet$`),               // .sheet(isPresented:) { ... }
-		regexp.MustCompile(`^fullScreenCover$`),     // .fullScreenCover(isPresented:) { ... }
-		regexp.MustCompile(`^popover$`),             // .popover(isPresented:) { ... }
-		regexp.MustCompile(`^alert$`),               // .alert(_:isPresented:) { ... }
-		regexp.MustCompile(`^confirmationDialog$`),  // .confirmationDialog(_:isPresented:) { ... }
-		regexp.MustCompile(`^fileImporter$`),        // .fileImporter(isPresented:)
-		regexp.MustCompile(`^fileExporter$`),        // .fileExporter(isPresented:)
-		regexp.MustCompile(`^listStyle$`),           // .listStyle(_:)
-		regexp.MustCompile(`^listRowInsets$`),       // .listRowInsets(_:)
-		regexp.MustCompile(`^listRowBackground$`),   // .listRowBackground(_:)
-		regexp.MustCompile(`^listRowSeparator$`),    // .listRowSeparator(_:)
-		regexp.MustCompile(`^navigationDestination$`), // .navigationDestination(for:) { ... }
-		regexp.MustCompile(`^tabItem$`),             // .tabItem { ... }
-		regexp.MustCompile(`^badge$`),               // .badge(_:)
-		regexp.MustCompile(`^tag$`),                 // .tag(_:)
-		regexp.MustCompile(`^id$`),                  // .id(_:) (identity modifier, not entity ID)
-		regexp.MustCompile(`^equatable$`),           // .equatable() (perf modifier)
-		regexp.MustCompile(`^drawingGroup$`),        // .drawingGroup()
-		regexp.MustCompile(`^compositingGroup$`),    // .compositingGroup()
-		regexp.MustCompile(`^contentShape$`),        // .contentShape(_:)
-		regexp.MustCompile(`^allowsHitTesting$`),    // .allowsHitTesting(_:)
-		regexp.MustCompile(`^accessibilityLabel$`),  // .accessibilityLabel(_:)
-		regexp.MustCompile(`^accessibilityHint$`),   // .accessibilityHint(_:)
-		regexp.MustCompile(`^accessibilityValue$`),  // .accessibilityValue(_:)
-		regexp.MustCompile(`^accessibilityHidden$`), // .accessibilityHidden(_:)
-		regexp.MustCompile(`^help$`),                // .help(_:) (accessibility tooltip)
-		regexp.MustCompile(`^disabled$`),            // .disabled(_:)
-		regexp.MustCompile(`^hidden$`),              // .hidden()
-		regexp.MustCompile(`^redacted$`),            // .redacted(reason:)
-		regexp.MustCompile(`^unredacted$`),          // .unredacted()
-		regexp.MustCompile(`^privacySensitive$`),    // .privacySensitive()
-		regexp.MustCompile(`^environment$`),         // .environment(_:_:)
-		regexp.MustCompile(`^environmentObject$`),   // .environmentObject(_:)
-		regexp.MustCompile(`^transformEnvironment$`), // .transformEnvironment(_:transform:)
-		regexp.MustCompile(`^preferenceKey$`),       // .preference(key:value:)
-		regexp.MustCompile(`^onPreferenceChange$`),  // .onPreferenceChange(_:) { ... }
-		regexp.MustCompile(`^transformPreference$`), // .transformPreference(_:_:)
-		regexp.MustCompile(`^anchorPreference$`),    // .anchorPreference(key:value:)
-		regexp.MustCompile(`^overlayPreferenceValue$`), // .overlayPreferenceValue(_:) { ... }
-		regexp.MustCompile(`^backgroundPreferenceValue$`), // .backgroundPreferenceValue(_:) { ... }
-		regexp.MustCompile(`^coordinateSpace$`),     // .coordinateSpace(name:)
-		regexp.MustCompile(`^matchedGeometryEffect$`), // .matchedGeometryEffect(id:in:)
-		regexp.MustCompile(`^edgesIgnoringSafeArea$`), // .edgesIgnoringSafeArea(_:)
-		regexp.MustCompile(`^ignoresSafeArea$`),     // .ignoresSafeArea(_:regions:)
-		regexp.MustCompile(`^safeAreaInset$`),       // .safeAreaInset(edge:content:)
-		regexp.MustCompile(`^clipped$`),             // .clipped()
-		regexp.MustCompile(`^mask$`),                // .mask(_:)
-		regexp.MustCompile(`^border$`),              // .border(_:width:)
-		regexp.MustCompile(`^imageScale$`),          // .imageScale(_:)
-		regexp.MustCompile(`^resizable$`),           // Image.resizable()
-		regexp.MustCompile(`^interpolation$`),       // Image.interpolation(_:)
-		regexp.MustCompile(`^antialiased$`),         // Image.antialiased(_:)
-		regexp.MustCompile(`^renderingMode$`),       // Image.renderingMode(_:)
-		regexp.MustCompile(`^symbolRenderingMode$`), // Image.symbolRenderingMode(_:)
-		regexp.MustCompile(`^symbolVariant$`),       // .symbolVariant(_:)
-		regexp.MustCompile(`^textCase$`),            // .textCase(_:)
-		regexp.MustCompile(`^textContentType$`),     // .textContentType(_:)
-		regexp.MustCompile(`^autocapitalization$`),  // .autocapitalization(_:)
-		regexp.MustCompile(`^autocorrectionDisabled$`), // .autocorrectionDisabled(_:)
-		regexp.MustCompile(`^keyboardType$`),        // .keyboardType(_:)
-		regexp.MustCompile(`^submitLabel$`),         // .submitLabel(_:)
-		regexp.MustCompile(`^lineLimit$`),           // .lineLimit(_:)
-		regexp.MustCompile(`^lineSpacing$`),         // .lineSpacing(_:)
-		regexp.MustCompile(`^multilineTextAlignment$`), // .multilineTextAlignment(_:)
-		regexp.MustCompile(`^truncationMode$`),      // .truncationMode(_:)
-		regexp.MustCompile(`^minimumScaleFactor$`),  // .minimumScaleFactor(_:)
-		regexp.MustCompile(`^allowsTightening$`),    // .allowsTightening(_:)
-		regexp.MustCompile(`^flipsForRightToLeftLayoutDirection$`), // .flipsForRightToLeftLayoutDirection(_:)
-		// UIKit methods that appear as bare leaf names when invoked on
-		// external UIKit types (UIButton, UIView, UIViewController, etc.).
-		regexp.MustCompile(`^addTarget$`),           // UIButton.addTarget(_:action:for:)
-		regexp.MustCompile(`^removeTarget$`),        // UIButton.removeTarget(_:action:for:)
-		regexp.MustCompile(`^addGestureRecognizer$`), // UIView.addGestureRecognizer(_:)
-		regexp.MustCompile(`^removeGestureRecognizer$`), // UIView.removeGestureRecognizer(_:)
-		regexp.MustCompile(`^addSubview$`),          // UIView.addSubview(_:)
-		regexp.MustCompile(`^removeFromSuperview$`), // UIView.removeFromSuperview()
-		regexp.MustCompile(`^insertSubview$`),       // UIView.insertSubview(_:at:)
-		regexp.MustCompile(`^bringSubviewToFront$`), // UIView.bringSubviewToFront(_:)
-		regexp.MustCompile(`^sendSubviewToBack$`),   // UIView.sendSubviewToBack(_:)
-		regexp.MustCompile(`^setNeedsLayout$`),      // UIView.setNeedsLayout()
-		regexp.MustCompile(`^setNeedsDisplay$`),     // UIView.setNeedsDisplay()
-		regexp.MustCompile(`^layoutIfNeeded$`),      // UIView.layoutIfNeeded()
-		regexp.MustCompile(`^layoutSubviews$`),      // UIView.layoutSubviews() override
-		regexp.MustCompile(`^sizeToFit$`),           // UIView.sizeToFit()
-		regexp.MustCompile(`^sizeThatFits$`),        // UIView.sizeThatFits(_:)
-		regexp.MustCompile(`^setNeedsUpdateConstraints$`), // UIView.setNeedsUpdateConstraints()
-		regexp.MustCompile(`^updateConstraints$`),   // UIView.updateConstraints() override
-		regexp.MustCompile(`^updateConstraintsIfNeeded$`), // UIView.updateConstraintsIfNeeded()
-		regexp.MustCompile(`^activate$`),            // NSLayoutConstraint.activate(_:)
-		regexp.MustCompile(`^deactivate$`),          // NSLayoutConstraint.deactivate(_:)
-		regexp.MustCompile(`^addConstraint$`),       // UIView.addConstraint(_:)
-		regexp.MustCompile(`^removeConstraint$`),    // UIView.removeConstraint(_:)
-		regexp.MustCompile(`^addConstraints$`),      // UIView.addConstraints(_:)
-		regexp.MustCompile(`^removeConstraints$`),   // UIView.removeConstraints(_:)
-		regexp.MustCompile(`^present$`),             // UIViewController.present(_:animated:)
-		regexp.MustCompile(`^dismiss$`),             // UIViewController.dismiss(animated:)
-		regexp.MustCompile(`^pushViewController$`),  // UINavigationController.pushViewController(_:animated:)
-		regexp.MustCompile(`^popViewController$`),   // UINavigationController.popViewController(animated:)
-		regexp.MustCompile(`^popToRootViewController$`), // UINavigationController.popToRootViewController(animated:)
-		regexp.MustCompile(`^performSegue$`),        // UIViewController.performSegue(withIdentifier:sender:)
-		regexp.MustCompile(`^reloadData$`),          // UITableView/UICollectionView.reloadData()
-		regexp.MustCompile(`^reloadRows$`),          // UITableView.reloadRows(at:with:)
-		regexp.MustCompile(`^reloadSections$`),      // UITableView.reloadSections(_:with:)
-		regexp.MustCompile(`^insertRows$`),          // UITableView.insertRows(at:with:)
-		regexp.MustCompile(`^deleteRows$`),          // UITableView.deleteRows(at:with:)
-		regexp.MustCompile(`^insertSections$`),      // UITableView.insertSections(_:with:)
-		regexp.MustCompile(`^deleteSections$`),      // UITableView.deleteSections(_:with:)
-		regexp.MustCompile(`^dequeueReusableCell$`), // UITableView.dequeueReusableCell(withIdentifier:)
-		regexp.MustCompile(`^dequeueReusableSupplementaryView$`), // UICollectionView
-		regexp.MustCompile(`^register$`),            // UITableView/UICollectionView.register(_:)
-		regexp.MustCompile(`^beginUpdates$`),        // UITableView.beginUpdates()
-		regexp.MustCompile(`^endUpdates$`),          // UITableView.endUpdates()
-		regexp.MustCompile(`^performBatchUpdates$`), // UICollectionView.performBatchUpdates(_:)
-		regexp.MustCompile(`^scrollToRow$`),         // UITableView.scrollToRow(at:at:animated:)
-		regexp.MustCompile(`^scrollToItem$`),        // UICollectionView.scrollToItem(at:at:animated:)
-		regexp.MustCompile(`^setEditing$`),          // UIViewController.setEditing(_:animated:)
-		regexp.MustCompile(`^becomeFirstResponder$`), // UIResponder.becomeFirstResponder()
-		regexp.MustCompile(`^resignFirstResponder$`), // UIResponder.resignFirstResponder()
-		regexp.MustCompile(`^endEditing$`),          // UIView.endEditing(_:)
-	}
-
-	// Elixir dynamic-pattern catalog (issue #44, slice 9).
-	//
-	// The Elixir extractor (internal/extractors/elixir/elixir.go) emits CALLS
-	// edges whose ToID is the bare trailing method identifier extracted from
-	// dotted call expressions — e.g. `Repo.all(User)` → ToID="all",
-	// `Phoenix.Controller.render(conn, ...)` → ToID="render". When the
-	// receiver is an external framework module (Ecto.Repo, Phoenix.Controller,
-	// Ecto.Changeset, GenServer behaviour, Logger) no in-tree entity exists
-	// for the callee, so the resolver cannot bind it. These stubs are
-	// statically unresolvable because:
-	//   1. The Ecto/Phoenix/OTP modules are imported from the Hex ecosystem
-	//      — they are never indexed as in-tree entities.
-	//   2. The extractor emits only the leaf name; without the receiver
-	//      module the resolver cannot distinguish `Repo.all` from a
-	//      user-defined function also named `all` on a domain module.
-	//
-	// The safer-bias rule (#94) is preserved by the per-language gate
-	// (lang=="elixir"): names like `all`, `get`, `insert`, `render`,
-	// `cast`, `change` are common in other ecosystems and must NOT fire
-	// for Go, Python, Ruby, TypeScript, etc.
-	//
-	// Four categories drive the bulk of unresolved Elixir edges:
-	//   1. Ecto.Repo query + mutation methods (`all`, `get`, `get!`, etc.)
-	//   2. Phoenix.Conn pipeline helpers (`render`, `json`, `send_resp`, etc.)
-	//   3. GenServer / OTP behaviour callbacks (`handle_call`, `handle_cast`, etc.)
-	//   4. Ecto.Changeset builder methods (`cast`, `validate_required`, etc.)
-	elixirDynamicPatterns = []*regexp.Regexp{
-		// ── Ecto.Repo query methods ──────────────────────────────────────
-		// `Repo.all/1`, `Repo.get/2`, `Repo.get!/2`, `Repo.get_by/2`,
-		// `Repo.get_by!/2`, `Repo.one/1`, `Repo.one!/1`, `Repo.aggregate/3`
-		// all arrive as bare leaf names because the extractor strips the
-		// receiver alias. No static resolver can bind them without full
-		// module-alias tracking.
-		regexp.MustCompile(`^all$`),
-		regexp.MustCompile(`^one$`),
-		regexp.MustCompile(`^get$`),
-		regexp.MustCompile(`^get!$`),
-		regexp.MustCompile(`^get_by$`),
-		regexp.MustCompile(`^get_by!$`),
-		regexp.MustCompile(`^aggregate$`),
-		regexp.MustCompile(`^preload$`),
-		regexp.MustCompile(`^reload$`),
-		regexp.MustCompile(`^transaction$`),
-		// Ecto.Repo mutation methods.
-		regexp.MustCompile(`^insert$`),
-		regexp.MustCompile(`^insert!$`),
-		regexp.MustCompile(`^insert_or_update$`),
-		regexp.MustCompile(`^insert_or_update!$`),
-		regexp.MustCompile(`^update$`),
-		regexp.MustCompile(`^update!$`),
-		regexp.MustCompile(`^delete$`),
-		regexp.MustCompile(`^delete!$`),
-		regexp.MustCompile(`^insert_all$`),
-		regexp.MustCompile(`^update_all$`),
-		regexp.MustCompile(`^delete_all$`),
-
-		// ── Phoenix.Conn pipeline helpers ───────────────────────────────
-		// Phoenix controller actions call bare helpers (render/json/
-		// send_resp) that are injected by `use Phoenix.Controller` via
-		// macro expansion — no static definition exists in the indexed tree.
-		regexp.MustCompile(`^render$`),
-		regexp.MustCompile(`^json$`),
-		regexp.MustCompile(`^text$`),
-		regexp.MustCompile(`^html$`),
-		regexp.MustCompile(`^send_resp$`),
-		regexp.MustCompile(`^send_file$`),
-		regexp.MustCompile(`^send_download$`),
-		regexp.MustCompile(`^put_flash$`),
-		regexp.MustCompile(`^clear_flash$`),
-		regexp.MustCompile(`^redirect$`),
-		regexp.MustCompile(`^halt$`),
-		regexp.MustCompile(`^assign$`),
-		regexp.MustCompile(`^fetch_session$`),
-		regexp.MustCompile(`^put_session$`),
-		regexp.MustCompile(`^get_session$`),
-		regexp.MustCompile(`^delete_session$`),
-		regexp.MustCompile(`^configure_session$`),
-		regexp.MustCompile(`^fetch_flash$`),
-		regexp.MustCompile(`^put_resp_content_type$`),
-		regexp.MustCompile(`^put_resp_header$`),
-		regexp.MustCompile(`^put_req_header$`),
-		regexp.MustCompile(`^delete_resp_header$`),
-		regexp.MustCompile(`^resp$`),
-
-		// ── GenServer / OTP behaviour callbacks ─────────────────────────
-		// OTP behaviour callbacks are invoked by the runtime, not by static
-		// call sites the resolver can see. The extractor emits them as
-		// ordinary CALLS stubs from the `init/start_link` entry point.
-		// snake_case + `handle_` prefix pattern is Elixir/Erlang-specific.
-		regexp.MustCompile(`^init$`),
-		regexp.MustCompile(`^handle_call$`),
-		regexp.MustCompile(`^handle_cast$`),
-		regexp.MustCompile(`^handle_info$`),
-		regexp.MustCompile(`^handle_continue$`),
-		regexp.MustCompile(`^terminate$`),
-		regexp.MustCompile(`^code_change$`),
-		// Agent / Task OTP helpers.
-		regexp.MustCompile(`^start_link$`),
-		regexp.MustCompile(`^start$`),
-		regexp.MustCompile(`^stop$`),
-		regexp.MustCompile(`^child_spec$`),
-		// GenServer reply helpers (Process, GenServer module-level calls).
-		regexp.MustCompile(`^reply$`),
-		regexp.MustCompile(`^noreply$`),
-
-		// ── Ecto.Changeset builder methods ──────────────────────────────
-		// Changeset pipelines (|> cast() |> validate_required() ...) arrive
-		// as bare leaf names. The `Ecto.Changeset` module is an external
-		// Hex dependency never indexed in-tree.
-		regexp.MustCompile(`^cast$`),
-		regexp.MustCompile(`^cast_assoc$`),
-		regexp.MustCompile(`^cast_embed$`),
-		regexp.MustCompile(`^validate_required$`),
-		regexp.MustCompile(`^validate_length$`),
-		regexp.MustCompile(`^validate_format$`),
-		regexp.MustCompile(`^validate_inclusion$`),
-		regexp.MustCompile(`^validate_exclusion$`),
-		regexp.MustCompile(`^validate_number$`),
-		regexp.MustCompile(`^validate_acceptance$`),
-		regexp.MustCompile(`^validate_confirmation$`),
-		regexp.MustCompile(`^validate_change$`),
-		regexp.MustCompile(`^validate_unique$`),
-		regexp.MustCompile(`^put_assoc$`),
-		regexp.MustCompile(`^put_embed$`),
-		regexp.MustCompile(`^put_change$`),
-		regexp.MustCompile(`^force_change$`),
-		regexp.MustCompile(`^change$`),
-		regexp.MustCompile(`^changeset$`),
-		regexp.MustCompile(`^add_error$`),
-		regexp.MustCompile(`^apply_action$`),
-		regexp.MustCompile(`^apply_action!$`),
-		regexp.MustCompile(`^apply_changes$`),
-		regexp.MustCompile(`^fetch_field$`),
-		regexp.MustCompile(`^fetch_field!$`),
-		regexp.MustCompile(`^get_field$`),
-		regexp.MustCompile(`^get_change$`),
-
-		// ── Logger module calls ──────────────────────────────────────────
-		// `Logger.debug/info/warn/error` arrive as bare `debug`/`info`/
-		// `warn`/`error` after receiver-stripping. Logger is an OTP built-in
-		// never indexed in-tree.
-		regexp.MustCompile(`^debug$`),
-		regexp.MustCompile(`^info$`),
-		regexp.MustCompile(`^warning$`),
-		regexp.MustCompile(`^error$`),
-		regexp.MustCompile(`^critical$`),
-		regexp.MustCompile(`^notice$`),
-
-		// ── Ecto.Query DSL (from / where / select / order_by …) ─────────
-		// `from q in Query, where: ..., select: ...` macro and named-
-		// binding query functions are injected by `import Ecto.Query`.
-		regexp.MustCompile(`^from$`),
-		regexp.MustCompile(`^where$`),
-		regexp.MustCompile(`^select$`),
-		regexp.MustCompile(`^order_by$`),
-		regexp.MustCompile(`^group_by$`),
-		regexp.MustCompile(`^having$`),
-		regexp.MustCompile(`^join$`),
-		regexp.MustCompile(`^limit$`),
-		regexp.MustCompile(`^offset$`),
-		regexp.MustCompile(`^lock$`),
-		regexp.MustCompile(`^distinct$`),
-		regexp.MustCompile(`^subquery$`),
-		regexp.MustCompile(`^exclude$`),
-		regexp.MustCompile(`^fragment$`),
-		regexp.MustCompile(`^dynamic$`),
-	}
-
-
-	// Zig dynamic-pattern catalog (issue #44 — Zig resolver slice).
-	//
-	// The Zig extractor (internal/extractors/zig/zig.go) emits CALLS edges
-	// whose ToID is the rightmost identifier in a dotted call chain.  For a
-	// stdlib call like `mem.splitScalar(u8, s, '/')` the receiver prefix is
-	// discarded and the stub is "splitScalar".  Many Zig std lib namespaces
-	// (std.mem, std.fmt, std.ascii, std.json, std.meta, std.ArrayList) expose
-	// function names that are entirely specific to Zig's standard library.
-	//
-	// Two categories:
-	//
-	//  1. Zig-unique std.mem / std.ascii / std.meta identifiers — names that
-	//     exist in no other language's standard library with the exact same
-	//     spelling and casing:
-	//       splitScalar, splitSequence, splitBackwardsScalar, tokenizeScalar
-	//       (std.mem split/tokenize family); indexOfScalar, lastIndexOfScalar
-	//       (std.mem search); eqlIgnoreCase, isPrint (std.ascii); stringToEnum
-	//       (std.meta); toOwnedSlice (std.ArrayList); parseFromSlice
-	//       (std.json); allocPrint (std.fmt).
-	//
-	//  2. Zig std lib leaf names gated to lang=="zig" — common enough words
-	//     that require the gate but are virtually always stdlib calls in real
-	//     Zig code:
-	//       stringify (std.json.stringify — safe; Go uses json.Marshal),
-	//       writeAll  (std.io Writer.writeAll — safe; Java uses write/println),
-	//       trim      (std.mem.trim — safe; Python .strip(), JS .trim() are
-	//                  method calls not bare functions in Zig context),
-	//       startsWith (std.mem.startsWith — safe; Python str.startswith() is
-	//                  a method, Zig exposes it as a bare module function),
-	//       parseInt   (std.fmt.parseInt — safe under gate; TypeScript
-	//                  parseInt is a global but Zig gate is tight enough).
-	//
-	// Excluded (too generic even under gate): assert, print, next, first, get,
-	// put, init, deinit, append, clamp, writer, entry, notFound, phrase.
-	//
-	// All patterns are gated to lang=="zig" (safer-bias rule #94).
-	zigDynamicPatterns = []*regexp.Regexp{
-		// ── Tier 1: Zig-unique std.mem split/tokenize family ────────────
-		// These camelCase compound names exist in no other language stdlib.
-		regexp.MustCompile(`^splitScalar$`),           // mem.splitScalar(T, slice, delimiter)
-		regexp.MustCompile(`^splitSequence$`),         // mem.splitSequence(T, slice, delimiter)
-		regexp.MustCompile(`^splitBackwardsScalar$`),  // mem.splitBackwardsScalar(T, slice, delimiter)
-		regexp.MustCompile(`^tokenizeScalar$`),        // mem.tokenizeScalar(T, slice, delimiter)
-		regexp.MustCompile(`^tokenizeSequence$`),      // mem.tokenizeSequence(T, slice, delimiter)
-		regexp.MustCompile(`^tokenizeAny$`),           // mem.tokenizeAny(T, slice, delimiter_bytes)
-
-		// ── Tier 1: Zig-unique std.mem search functions ─────────────────
-		regexp.MustCompile(`^indexOfScalar$`),         // mem.indexOfScalar(T, slice, value)
-		regexp.MustCompile(`^lastIndexOfScalar$`),     // mem.lastIndexOfScalar(T, slice, value)
-		regexp.MustCompile(`^indexOfPosLinear$`),      // mem.indexOfPosLinear(T, slice, pos, needle)
-		regexp.MustCompile(`^indexOfPos$`),            // mem.indexOfPos(T, slice, start, needle)
-		regexp.MustCompile(`^lastIndexOf$`),           // mem.lastIndexOf(T, slice, value) — under Zig gate
-		regexp.MustCompile(`^containsAtLeast$`),       // mem.containsAtLeast(T, slice, count, value)
-
-		// ── Tier 1: Zig-unique std.ascii functions ───────────────────────
-		regexp.MustCompile(`^eqlIgnoreCase$`),         // ascii.eqlIgnoreCase(a, b)
-		regexp.MustCompile(`^isPrint$`),               // ascii.isPrint(c)
-		regexp.MustCompile(`^isAlpha$`),               // ascii.isAlpha(c)
-		regexp.MustCompile(`^isAlNum$`),               // ascii.isAlNum(c)
-		regexp.MustCompile(`^isDigit$`),               // ascii.isDigit(c)
-		regexp.MustCompile(`^isUpper$`),               // ascii.isUpper(c)
-		regexp.MustCompile(`^isLower$`),               // ascii.isLower(c)
-		regexp.MustCompile(`^toUpper$`),               // ascii.toUpper(c)
-		regexp.MustCompile(`^toLower$`),               // ascii.toLower(c)
-		regexp.MustCompile(`^lowerString$`),           // ascii.lowerString(buffer, input)
-		regexp.MustCompile(`^upperString$`),           // ascii.upperString(buffer, input)
-
-		// ── Tier 1: Zig-unique std.meta / std.fmt / std.json ────────────
-		regexp.MustCompile(`^stringToEnum$`),          // meta.stringToEnum(T, str)
-		regexp.MustCompile(`^toOwnedSlice$`),          // ArrayList.toOwnedSlice()
-		regexp.MustCompile(`^parseFromSlice$`),        // json.parseFromSlice(T, alloc, slice, opts)
-		regexp.MustCompile(`^allocPrint$`),            // fmt.allocPrint(alloc, template, args)
-		regexp.MustCompile(`^allocPrintZ$`),           // fmt.allocPrintZ (null-terminated variant)
-		regexp.MustCompile(`^bufPrint$`),              // fmt.bufPrint(buf, template, args)
-		regexp.MustCompile(`^parseInt$`),              // fmt.parseInt(T, str, radix)
-
-		// ── Tier 2: Zig std lib names gated to lang=="zig" ──────────────
-		// Common words but safe under the Zig gate (real Zig code only uses
-		// them as std module calls, never as user-defined method names).
-		regexp.MustCompile(`^stringify$`),             // json.stringify(value, opts, writer)
-		regexp.MustCompile(`^writeAll$`),              // Writer.writeAll(slice)
-		regexp.MustCompile(`^startsWith$`),            // mem.startsWith(T, slice, prefix)
-		regexp.MustCompile(`^endsWith$`),              // mem.endsWith(T, slice, suffix)
-		regexp.MustCompile(`^trim$`),                  // mem.trim(T, slice, chars)
-		regexp.MustCompile(`^trimLeft$`),              // mem.trimLeft(T, slice, chars)
-		regexp.MustCompile(`^trimRight$`),             // mem.trimRight(T, slice, chars)
-		regexp.MustCompile(`^eql$`),                   // mem.eql(T, a, b) — Zig gate: Go uses bytes.Equal
-		regexp.MustCompile(`^replaceOwned$`),          // mem.replaceOwned(T, alloc, input, needle, replacement)
-		regexp.MustCompile(`^concat$`),                // mem.concat(T, alloc, slices) — Zig gate; Lua already gated
-		// Zig std.io.Writer extended methods — Zig-specific vectored I/O.
-		regexp.MustCompile(`^writeVecAll$`),           // Writer.writeVecAll(vecs) vectored write
-		regexp.MustCompile(`^writeBytesNTimes$`),      // Writer.writeBytesNTimes(bytes, n)
-		// Zig std.mem fill/copy functions.
-		regexp.MustCompile(`^copyForwards$`),          // mem.copyForwards(T, dest, src)
-		regexp.MustCompile(`^copyBackwards$`),         // mem.copyBackwards(T, dest, src)
-		regexp.MustCompile(`^zeroes$`),                // mem.zeroes(T)
-		// Zig std.fmt number-to-string helpers.
-		regexp.MustCompile(`^formatInt$`),             // fmt.formatInt(value, radix, case, buf)
-		regexp.MustCompile(`^formatIntBuf$`),          // fmt.formatIntBuf(buf, value, radix, case, fill)
-		regexp.MustCompile(`^parseFloat$`),            // fmt.parseFloat(T, buf)
-		// Zig heap/allocator helpers.
-		regexp.MustCompile(`^ArenaAllocator$`),        // std.heap.ArenaAllocator — type constructor call
-		regexp.MustCompile(`^allocator$`),             // ArenaAllocator.allocator()
-		regexp.MustCompile(`^toOwnedSliceSentinel$`),  // ArrayList.toOwnedSliceSentinel(0)
-		// std.debug helpers — gated to Zig so they don't match Go's testing.T
-		// or Python's assert statements (which are keywords, not function calls).
-		regexp.MustCompile(`^print$`),                 // std.debug.print(fmt, args) — extremely common
-		regexp.MustCompile(`^panic$`),                 // std.debug.panic(msg) — Zig-gate: Go uses panic() too
-		regexp.MustCompile(`^assert$`),                // std.debug.assert(cond) — Zig uses it as fn call
-		// std.math helpers — clamp and abs exist in other stdlibs but under
-		// the Zig gate they map uniquely to std.math.{clamp,abs}.
-		regexp.MustCompile(`^clamp$`),                 // math.clamp(val, min, max)
-		regexp.MustCompile(`^isNan$`),                 // math.isNan(v)
-		regexp.MustCompile(`^isInf$`),                 // math.isInf(v, sign)
-		regexp.MustCompile(`^log2$`),                  // math.log2(v)
-		regexp.MustCompile(`^log10$`),                 // math.log10(v)
-		// std.mem — additional patterns not covered above.
-		regexp.MustCompile(`^bytesAsSlice$`),          // mem.bytesAsSlice(T, bytes)
-		regexp.MustCompile(`^sliceAsBytes$`),          // mem.sliceAsBytes(slice)
-		regexp.MustCompile(`^asBytes$`),               // mem.asBytes(ptr)
-		regexp.MustCompile(`^readInt$`),               // mem.readInt(T, bytes, endian)
-		regexp.MustCompile(`^writeInt$`),              // mem.writeInt(T, bytes, value, endian)
-		regexp.MustCompile(`^writeIntNative$`),        // mem.writeIntNative(T, bytes, value)
-	}
-
-	// Cross-language patterns that are safe to evaluate when language is
-	// unknown. Template-built names (`${x}` interpolation) are reflection-
+	// crossLangDynamicPatterns are patterns safe for any language (e.g.
+	// template-built names like `${x}`). crossLangDynamicPatterns is
+	// checked when the language is unknown or when no per-language slice
+	// matches the source language.
+	// Patterns that are intrinsically reflective regardless of language
+	// (template-built names like `${x}` interpolation) are reflection-
 	// shaped in every language that has them.
 	crossLangDynamicPatterns = []*regexp.Regexp{
 		regexp.MustCompile(`.*\$\{.*\}.*`), // template-built strings ${x}
 	}
 
-	// C# / .NET dynamic-pattern catalog (issue #441). Routes
-	// project-internal `using Namespace.SubNamespace;` IMPORTS whose
-	// target dotted namespace has no entity in the graph (because we
-	// index file-level entities, not namespace entities) into Dynamic
-	// instead of bug-extractor. Pattern: PascalCase root segment, at
-	// least one dot, no leading `Microsoft.`/`System.` (those resolve
-	// via the external allowlist) and no leading lowercase (which would
-	// be a method-on-receiver shape, handled elsewhere).
-	//
-	// Also covers a small set of receiver-stripped reflection /
-	// concurrency primitives the C# extractor emits as bare or dotted
-	// callees (`Interlocked.Increment`, `MethodBase.GetCurrentMethod`,
-	// `PeriodicTimer.WaitForNextTickAsync`, `ConcurrentDictionary.
-	// TryRemove`). These are not language-builtins in the strict sense
-	// but they're framework-dispatch entry points that no static binder
-	// can reach without full assembly-level type resolution.
-	csharpDynamicPatterns = []*regexp.Regexp{
-		// PascalCase project-internal namespace import.
-		// Anchored: starts with uppercase, contains at least one dot, every
-		// segment is an identifier (no whitespace / brackets), and there is
-		// no generic `<...>` suffix (those are call sites, not imports).
-		// Negative-lookahead would be cleaner but Go regexp's RE2 dialect
-		// doesn't support lookaround — we filter out the well-known .NET
-		// / Microsoft / EF Core ecosystem roots at the caller (handled in
-		// isDynamicPatternLang via a startsWith check before regex eval).
-		regexp.MustCompile(`^[A-Z][A-Za-z0-9_]*(?:\.[A-Z][A-Za-z0-9_]*)+$`),
-
-		// Issue #44 — Quartz.NET / Hangfire generic static-factory calls.
-		// The C# extractor emits `JobBuilder.Create<ReportJob>` and
-		// `BackgroundJob.Enqueue<IEmailService>` as bare CALLS stubs when the
-		// receiver type carries a generic type argument. The existing PascalCase-
-		// dotted pattern above only matches stubs without a `<` suffix; these
-		// generic forms fall through to BugExtractor. Pattern: PascalCase
-		// dotted identifier immediately followed by `<` (the opening of the
-		// type-argument list). C#-language gate keeps this from firing on
-		// non-.NET codebases (Go generics, TypeScript generics, etc.).
-		regexp.MustCompile(`^[A-Z][A-Za-z0-9_]*\.[A-Z][A-Za-z0-9_]*<`),
-
-		// Issue #44 — Quartz.NET fluent builder bare-name leaf methods.
-		// The Quartz.NET JobBuilder / TriggerBuilder API uses a fluent chain:
-		//   JobBuilder.Create<T>().WithIdentity("name").Build()
-		//   TriggerBuilder.Create().WithIdentity("t").StartNow().Build()
-		// After the extractor strips the receiver the bare method leaves
-		// `WithIdentity` and `StartNow` land in BugExtractor because they
-		// are not in any allowlist. Without full type inference the resolver
-		// cannot bind these to the Quartz.NET builder — Dynamic is the
-		// correct disposition. C#-language gate (#94 safer-bias rule) keeps
-		// these from polluting non-.NET graphs.
-		regexp.MustCompile(`^WithIdentity$`), // JobBuilder/TriggerBuilder.WithIdentity(...)
-		regexp.MustCompile(`^StartNow$`),     // TriggerBuilder.StartNow()
-	}
-
-	// csharpExternalNamespaceRoots lists the dotted namespace roots that
-	// must NOT be classified as Dynamic by csharpDynamicPatterns — they
-	// are real external imports (Microsoft.AspNetCore, System.Linq, EF
-	// Core, etc.) that the external synthesiser routes to ext:microsoft
-	// / ext:system. Without this exclusion the dynamic-pattern check
-	// (which runs BEFORE the external-prefix check, Refs #95) would
-	// promote every Microsoft.* import to Dynamic and the corpus would
-	// lose its ExternalKnown classification.
-	csharpExternalNamespaceRoots = []string{
-		"Microsoft.",
-		"System.",
-		"EntityFrameworkCore",
-		"Newtonsoft.",
-		"Serilog",
-		"NLog",
-		"Autofac",
-		"Castle.",
-		"AutoMapper",
-		"MediatR",
-		"FluentValidation",
-		"FluentAssertions",
-		"NUnit",
-		"Xunit",
-		"Moq",
-		"Polly",
-		"Dapper",
-		"RestSharp",
-		"Hangfire",
-		"Quartz",
-		"IdentityServer",
-		"MassTransit",
-		"NServiceBus",
-		"RabbitMQ.",
-		"StackExchange.",
-		"Swashbuckle.",
-		"GraphQL.",
-		"HotChocolate",
-		"AspNetCore",
-		"Org.BouncyCastle",
-		"Mvc.",
-	}
-
-	// dartDynamicPatterns covers three categories of statically-unresolvable
-	// call sites in Flutter / Dart code:
-	//
-	//  1. Package-URI import ToIDs — the Dart extractor emits the full import
-	//     URI as the IMPORTS edge ToID (e.g. `package:flutter/material.dart`,
-	//     `dart:convert`, `src/widgets/cart.dart`). These URIs are not
-	//     in-tree entity names; the resolver cannot map them to a graph entity
-	//     without a package-resolution pass. Route to Dynamic (same precedent
-	//     as Python relative imports and TS path-mapped imports).
-	//
-	//  2. Flutter/Material widget constructor calls — the Dart extractor
-	//     strips the receiver from `Widget(...)` expressions and emits only
-	//     the constructor name as the CALLS ToID. Flutter widgets are
-	//     framework types imported from `package:flutter/material.dart` (or
-	//     cupertino / widgets packages) and have no in-tree entity. The most
-	//     frequent hits are enumerated below; the per-language gate keeps
-	//     these names from shadowing user-defined classes in other languages.
-	//
-	//  3. ChangeNotifier / StatefulWidget lifecycle and async/Future chain
-	//     methods — `notifyListeners`, `setState`, `then`, `catchError`,
-	//     `whenComplete`, `listen` are called on framework-provided or
-	//     Dart-stdlib objects; the receiver is stripped by the extractor and
-	//     static binding is impossible without full type inference.
-	dartDynamicPatterns = []*regexp.Regexp{
-		// ── 1. Package-URI import ToIDs ──────────────────────────────────
-		// `package:` imports: `package:flutter/material.dart`,
-		// `package:provider/provider.dart`, `package:http/http.dart`, etc.
-		regexp.MustCompile(`^package:`),
-		// `dart:` standard-library imports: `dart:convert`, `dart:async`, etc.
-		regexp.MustCompile(`^dart:`),
-		// Relative path imports: `../widgets/cart.dart`, `./util.dart`,
-		// `src/screens/home.dart`. Same shape as the JS/TS relative-import
-		// rule — one SCOPE.Component placeholder per importing file, so
-		// bare-name lookup is ambiguous.
-		regexp.MustCompile(`^\.{1,2}/`),
-		regexp.MustCompile(`^\.{1,3}$`),
-		regexp.MustCompile(`^(lib|src|test|integration_test)/`),
-
-		// ── 2. Flutter/Material/Cupertino widget constructors ────────────
-		// Layout widgets
-		regexp.MustCompile(`^Column$`),
-		regexp.MustCompile(`^Row$`),
-		regexp.MustCompile(`^Stack$`),
-		regexp.MustCompile(`^Wrap$`),
-		regexp.MustCompile(`^Flex$`),
-		regexp.MustCompile(`^Expanded$`),
-		regexp.MustCompile(`^Flexible$`),
-		regexp.MustCompile(`^Spacer$`),
-		regexp.MustCompile(`^SizedBox$`),
-		regexp.MustCompile(`^ConstrainedBox$`),
-		regexp.MustCompile(`^FittedBox$`),
-		regexp.MustCompile(`^LimitedBox$`),
-		regexp.MustCompile(`^OverflowBox$`),
-		regexp.MustCompile(`^UnconstrainedBox$`),
-		regexp.MustCompile(`^AspectRatio$`),
-		regexp.MustCompile(`^Center$`),
-		regexp.MustCompile(`^Align$`),
-		regexp.MustCompile(`^Padding$`),
-		regexp.MustCompile(`^Container$`),
-		regexp.MustCompile(`^DecoratedBox$`),
-		regexp.MustCompile(`^Positioned$`),
-		regexp.MustCompile(`^Expanded$`),
-		// Scrolling
-		regexp.MustCompile(`^ListView$`),
-		regexp.MustCompile(`^GridView$`),
-		regexp.MustCompile(`^SingleChildScrollView$`),
-		regexp.MustCompile(`^CustomScrollView$`),
-		regexp.MustCompile(`^PageView$`),
-		regexp.MustCompile(`^NestedScrollView$`),
-		regexp.MustCompile(`^RefreshIndicator$`),
-		// Sliver delegates (common SliverGrid / SliverList)
-		regexp.MustCompile(`^SliverList$`),
-		regexp.MustCompile(`^SliverGrid$`),
-		regexp.MustCompile(`^SliverGridDelegateWith(?:Fixed|Max)CrossAxisCount$`),
-		regexp.MustCompile(`^SliverChildListDelegate$`),
-		regexp.MustCompile(`^SliverChildBuilderDelegate$`),
-		// Scaffold & navigation structure
-		regexp.MustCompile(`^Scaffold$`),
-		regexp.MustCompile(`^AppBar$`),
-		regexp.MustCompile(`^BottomNavigationBar$`),
-		regexp.MustCompile(`^BottomNavigationBarItem$`),
-		regexp.MustCompile(`^NavigationBar$`),
-		regexp.MustCompile(`^NavigationRail$`),
-		regexp.MustCompile(`^Drawer$`),
-		regexp.MustCompile(`^DrawerHeader$`),
-		regexp.MustCompile(`^TabBar$`),
-		regexp.MustCompile(`^TabBarView$`),
-		regexp.MustCompile(`^TabController$`),
-		regexp.MustCompile(`^DefaultTabController$`),
-		// App-level
-		regexp.MustCompile(`^MaterialApp$`),
-		regexp.MustCompile(`^WidgetsApp$`),
-		regexp.MustCompile(`^CupertinoApp$`),
-		regexp.MustCompile(`^MultiProvider$`),
-		regexp.MustCompile(`^ChangeNotifierProvider$`),
-		regexp.MustCompile(`^Provider$`),
-		regexp.MustCompile(`^Consumer$`),
-		regexp.MustCompile(`^Selector$`),
-		// Text & typography
-		regexp.MustCompile(`^Text$`),
-		regexp.MustCompile(`^RichText$`),
-		regexp.MustCompile(`^SelectableText$`),
-		regexp.MustCompile(`^TextSpan$`),
-		regexp.MustCompile(`^DefaultTextStyle$`),
-		// Buttons
-		regexp.MustCompile(`^ElevatedButton$`),
-		regexp.MustCompile(`^TextButton$`),
-		regexp.MustCompile(`^OutlinedButton$`),
-		regexp.MustCompile(`^FilledButton$`),
-		regexp.MustCompile(`^IconButton$`),
-		regexp.MustCompile(`^FloatingActionButton$`),
-		regexp.MustCompile(`^DropdownButton$`),
-		regexp.MustCompile(`^DropdownMenuItem$`),
-		regexp.MustCompile(`^PopupMenuButton$`),
-		regexp.MustCompile(`^PopupMenuItem$`),
-		// Forms & input
-		regexp.MustCompile(`^TextField$`),
-		regexp.MustCompile(`^TextFormField$`),
-		regexp.MustCompile(`^Form$`),
-		regexp.MustCompile(`^FormField$`),
-		regexp.MustCompile(`^Checkbox$`),
-		regexp.MustCompile(`^CheckboxListTile$`),
-		regexp.MustCompile(`^Radio$`),
-		regexp.MustCompile(`^RadioListTile$`),
-		regexp.MustCompile(`^Switch$`),
-		regexp.MustCompile(`^SwitchListTile$`),
-		regexp.MustCompile(`^Slider$`),
-		regexp.MustCompile(`^RangeSlider$`),
-		regexp.MustCompile(`^DatePicker$`),
-		regexp.MustCompile(`^TimePicker$`),
-		regexp.MustCompile(`^InputDecoration$`),
-		// Decoration & theming
-		regexp.MustCompile(`^BoxDecoration$`),
-		regexp.MustCompile(`^ShapeDecoration$`),
-		regexp.MustCompile(`^BoxConstraints$`),
-		regexp.MustCompile(`^BorderRadius$`),
-		regexp.MustCompile(`^RoundedRectangleBorder$`),
-		regexp.MustCompile(`^Border$`),
-		regexp.MustCompile(`^BorderSide$`),
-		regexp.MustCompile(`^ThemeData$`),
-		regexp.MustCompile(`^Theme$`),
-		regexp.MustCompile(`^ColorScheme$`),
-		regexp.MustCompile(`^TextStyle$`),
-		regexp.MustCompile(`^TextTheme$`),
-		regexp.MustCompile(`^EdgeInsets$`),
-		regexp.MustCompile(`^EdgeInsetsDirectional$`),
-		regexp.MustCompile(`^Gradient$`),
-		regexp.MustCompile(`^LinearGradient$`),
-		regexp.MustCompile(`^RadialGradient$`),
-		regexp.MustCompile(`^SweepGradient$`),
-		regexp.MustCompile(`^BoxShadow$`),
-		// Images & icons
-		regexp.MustCompile(`^Image$`),
-		regexp.MustCompile(`^Icon$`),
-		regexp.MustCompile(`^CircleAvatar$`),
-		regexp.MustCompile(`^ClipRRect$`),
-		regexp.MustCompile(`^ClipOval$`),
-		regexp.MustCompile(`^ClipPath$`),
-		// Lists & cards
-		regexp.MustCompile(`^Card$`),
-		regexp.MustCompile(`^ListTile$`),
-		regexp.MustCompile(`^ExpansionTile$`),
-		regexp.MustCompile(`^Divider$`),
-		// Overlays & dialogs
-		regexp.MustCompile(`^AlertDialog$`),
-		regexp.MustCompile(`^SimpleDialog$`),
-		regexp.MustCompile(`^BottomSheet$`),
-		regexp.MustCompile(`^SnackBar$`),
-		regexp.MustCompile(`^SnackBarAction$`),
-		regexp.MustCompile(`^Tooltip$`),
-		regexp.MustCompile(`^Banner$`),
-		// Progress & loading
-		regexp.MustCompile(`^CircularProgressIndicator$`),
-		regexp.MustCompile(`^LinearProgressIndicator$`),
-		// Gesture & interaction
-		regexp.MustCompile(`^GestureDetector$`),
-		regexp.MustCompile(`^InkWell$`),
-		regexp.MustCompile(`^InkResponse$`),
-		regexp.MustCompile(`^Draggable$`),
-		regexp.MustCompile(`^DragTarget$`),
-		// Safe area, media, focus
-		regexp.MustCompile(`^SafeArea$`),
-		regexp.MustCompile(`^MediaQuery$`),
-		regexp.MustCompile(`^FocusScope$`),
-		regexp.MustCompile(`^Focus$`),
-		// Builder / async widgets
-		regexp.MustCompile(`^Builder$`),
-		regexp.MustCompile(`^StatefulBuilder$`),
-		regexp.MustCompile(`^FutureBuilder$`),
-		regexp.MustCompile(`^StreamBuilder$`),
-		regexp.MustCompile(`^AnimatedBuilder$`),
-		regexp.MustCompile(`^ValueListenableBuilder$`),
-		regexp.MustCompile(`^AnimatedWidget$`),
-		// Navigation
-		regexp.MustCompile(`^MaterialPageRoute$`),
-		regexp.MustCompile(`^CupertinoPageRoute$`),
-		regexp.MustCompile(`^PageRouteBuilder$`),
-		// Misc Material
-		regexp.MustCompile(`^Chip$`),
-		regexp.MustCompile(`^Badge$`),
-		regexp.MustCompile(`^Stepper$`),
-		regexp.MustCompile(`^DataTable$`),
-		regexp.MustCompile(`^DataRow$`),
-		regexp.MustCompile(`^DataCell$`),
-		regexp.MustCompile(`^DataColumn$`),
-
-		// ── 3. ChangeNotifier / StatefulWidget lifecycle ─────────────────
-		// `notifyListeners()` — ChangeNotifier method; receiver is `this`
-		// (stripped by extractor). Not user-defined, not resolvable.
-		regexp.MustCompile(`^notifyListeners$`),
-		// `setState(() { ... })` — StatefulWidget lifecycle; same pattern.
-		regexp.MustCompile(`^setState$`),
-		// `addListener` / `removeListener` — Listenable API.
-		regexp.MustCompile(`^addListener$`),
-		regexp.MustCompile(`^removeListener$`),
-		// `markNeedsBuild` / `markNeedsPaint` — internal widget tree methods.
-		regexp.MustCompile(`^markNeedsBuild$`),
-		regexp.MustCompile(`^markNeedsPaint$`),
-		// `initState` / `dispose` / `didUpdateWidget` / `didChangeDependencies`
-		// appear on CONTAINS edges when a State override calls super.*; they
-		// also appear as bare CALLS inside override bodies when calling
-		// `super.initState()` (super receiver stripped to `initState`).
-		regexp.MustCompile(`^initState$`),
-		regexp.MustCompile(`^didUpdateWidget$`),
-		regexp.MustCompile(`^didChangeDependencies$`),
-		regexp.MustCompile(`^reassemble$`),
-		regexp.MustCompile(`^deactivate$`),
-
-		// ── 4. Dart async / Future / Stream chain methods ────────────────
-		// These arrive as bare identifiers after the extractor strips the
-		// receiver (e.g. `future.then(...)` → ToID="then"). The actual
-		// implementation lives in dart:async; static binding is impossible.
-		// Dart-only gate prevents collision with Ruby's `then` (real method)
-		// and TS `then`/`catch`/`finally` (already in jsDynamicPatterns).
-		regexp.MustCompile(`^catchError$`),
-		regexp.MustCompile(`^whenComplete$`),
-		regexp.MustCompile(`^listen$`),
-		regexp.MustCompile(`^asBroadcastStream$`),
-		regexp.MustCompile(`^asyncExpand$`),
-		regexp.MustCompile(`^asyncMap$`),
-		regexp.MustCompile(`^onError$`),
-
-		// ── 5. Dart core / stdlib method names stripped from receivers ───
-		// `of` — static factory ubiquitous in Flutter: `Theme.of(context)`,
-		// `Navigator.of(context)`, `Provider.of<T>(context)`,
-		// `ScaffoldMessenger.of(context)`, `MediaQuery.of(context)`.
-		// The receiver is stripped by the extractor; `of` alone is
-		// unresolvable. Dart-only gate: `of` in Python/Go/Ruby/Java is
-		// a real user method name.
-		regexp.MustCompile(`^of$`),
-		// `maybeOf` — nullable variant of the static factory pattern.
-		regexp.MustCompile(`^maybeOf$`),
-		// `showSnackBar` / `showDialog` / `showModalBottomSheet` —
-		// called on ScaffoldMessenger.of(context) / showDialog(context:...)
-		// after receiver strip.
-		regexp.MustCompile(`^showSnackBar$`),
-		regexp.MustCompile(`^showDialog$`),
-		regexp.MustCompile(`^showModalBottomSheet$`),
-		regexp.MustCompile(`^showBottomSheet$`),
-		regexp.MustCompile(`^showDatePicker$`),
-		regexp.MustCompile(`^showTimePicker$`),
-		regexp.MustCompile(`^showSearch$`),
-		// Navigator push/pop — called on `Navigator.of(context)` after strip.
-		regexp.MustCompile(`^pushNamed$`),
-		regexp.MustCompile(`^pushReplacementNamed$`),
-		regexp.MustCompile(`^push$`),
-		regexp.MustCompile(`^pushReplacement$`),
-		regexp.MustCompile(`^pop$`),
-		regexp.MustCompile(`^popAndPushNamed$`),
-		regexp.MustCompile(`^canPop$`),
-		// `mounted` check — bool getter on State<T>; not a call but can
-		// appear as a bare reference after control-flow extraction.
-		// Not applicable as a call stub; omitted.
-
-		// ── 6. Dart core / dart:core type constructors and methods ───────
-		// `Exception(...)` — Dart built-in exception constructor; the
-		// extractor strips `throw Exception(...)` to bare `Exception`.
-		regexp.MustCompile(`^Exception$`),
-		regexp.MustCompile(`^FormatException$`),
-		regexp.MustCompile(`^ArgumentError$`),
-		regexp.MustCompile(`^StateError$`),
-		regexp.MustCompile(`^RangeError$`),
-		regexp.MustCompile(`^UnsupportedError$`),
-		regexp.MustCompile(`^UnimplementedError$`),
-		regexp.MustCompile(`^TypeError$`),
-		regexp.MustCompile(`^AssertionError$`),
-		regexp.MustCompile(`^Error$`),
-		// `toString()` — dart:core Object method; every Dart object has it.
-		// Dart-gate prevents collision with Go `.String()` helper / Python
-		// `str()` / Java `toString()` (those are handled by their own
-		// per-language logic or not present in the dynamic catalog).
-		regexp.MustCompile(`^toString$`),
-		// `parse` / `tryParse` — static factory on int/double/DateTime/Uri.
-		// The receiver is stripped by the extractor; the bare `parse` callee
-		// cannot be resolved without knowing the receiver type.
-		regexp.MustCompile(`^parse$`),
-		regexp.MustCompile(`^tryParse$`),
-		// `decode` / `encode` — json.decode / json.encode / utf8.decode
-		// stripped of the receiver; dart:convert stdlib, not user code.
-		regexp.MustCompile(`^decode$`),
-		regexp.MustCompile(`^encode$`),
-		// `now` — DateTime.now() static factory.
-		regexp.MustCompile(`^now$`),
-		// `fromSeed` — ColorScheme.fromSeed() static factory (Material 3).
-		regexp.MustCompile(`^fromSeed$`),
-		// `circular` — BorderRadius.circular() static factory.
-		regexp.MustCompile(`^circular$`),
-		// `network` — Image.network() constructor.
-		regexp.MustCompile(`^network$`),
-		// `icon` — ElevatedButton.icon(), OutlinedButton.icon() named constructors.
-		regexp.MustCompile(`^icon$`),
-		// `builder` — GridView.builder(), ListView.builder() named constructors.
-		regexp.MustCompile(`^builder$`),
-		// `all` — EdgeInsets.all(), SizedBox.all(), BoxConstraints.expand()
-		// pattern; most common: EdgeInsets.all().
-		regexp.MustCompile(`^all$`),
-		// `only` / `symmetric` — EdgeInsets.only() / EdgeInsets.symmetric().
-		regexp.MustCompile(`^only$`),
-		regexp.MustCompile(`^symmetric$`),
-
-		// ── 7. dart:collection / Dart Map + Iterable stdlib methods ──────
-		// Called on Map<K,V> / List<T> / Set<T> receivers that the
-		// extractor strips to bare identifiers. These are dart:core stdlib
-		// methods; no in-tree entity exists for them.
-		regexp.MustCompile(`^containsKey$`),
-		regexp.MustCompile(`^containsValue$`),
-		regexp.MustCompile(`^putIfAbsent$`),
-		regexp.MustCompile(`^remove$`),
-		regexp.MustCompile(`^update$`),
-		regexp.MustCompile(`^toList$`),
-		regexp.MustCompile(`^toSet$`),
-		regexp.MustCompile(`^toMap$`),
-		regexp.MustCompile(`^forEach$`),
-		regexp.MustCompile(`^map$`),
-		regexp.MustCompile(`^where$`),
-		regexp.MustCompile(`^reduce$`),
-		regexp.MustCompile(`^fold$`),
-		regexp.MustCompile(`^any$`),
-		regexp.MustCompile(`^every$`),
-		regexp.MustCompile(`^firstWhere$`),
-		regexp.MustCompile(`^lastWhere$`),
-		regexp.MustCompile(`^singleWhere$`),
-		regexp.MustCompile(`^expand$`),
-		regexp.MustCompile(`^skip$`),
-		regexp.MustCompile(`^take$`),
-		regexp.MustCompile(`^sort$`),
-		regexp.MustCompile(`^reversed$`),
-		regexp.MustCompile(`^isEmpty$`),
-		regexp.MustCompile(`^isNotEmpty$`),
-
-		// ── 8. dart:core String methods ───────────────────────────────────
-		regexp.MustCompile(`^trim$`),
-		regexp.MustCompile(`^trimLeft$`),
-		regexp.MustCompile(`^trimRight$`),
-		regexp.MustCompile(`^split$`),
-		regexp.MustCompile(`^join$`),
-		regexp.MustCompile(`^toLowerCase$`),
-		regexp.MustCompile(`^toUpperCase$`),
-		regexp.MustCompile(`^contains$`),
-		regexp.MustCompile(`^startsWith$`),
-		regexp.MustCompile(`^endsWith$`),
-		regexp.MustCompile(`^replaceAll$`),
-		regexp.MustCompile(`^replaceFirst$`),
-		regexp.MustCompile(`^indexOf$`),
-		regexp.MustCompile(`^lastIndexOf$`),
-		regexp.MustCompile(`^substring$`),
-		regexp.MustCompile(`^padLeft$`),
-		regexp.MustCompile(`^padRight$`),
-		// `toStringAsFixed` / `toStringAsPrecision` — double/num methods.
-		regexp.MustCompile(`^toStringAsFixed$`),
-		regexp.MustCompile(`^toStringAsPrecision$`),
-		regexp.MustCompile(`^toStringAsExponential$`),
-		regexp.MustCompile(`^toDouble$`),
-		regexp.MustCompile(`^toInt$`),
-		regexp.MustCompile(`^abs$`),
-		regexp.MustCompile(`^ceil$`),
-		regexp.MustCompile(`^floor$`),
-		regexp.MustCompile(`^round$`),
-		// `validate` — GlobalKey<FormState>.currentState!.validate().
-		regexp.MustCompile(`^validate$`),
-		regexp.MustCompile(`^reset$`),
-		regexp.MustCompile(`^save$`),
-		// `runApp` — Flutter framework entry point (dart:ui / package:flutter).
-		regexp.MustCompile(`^runApp$`),
-		// `dispose` — TextEditingController.dispose() / ScrollController.dispose()
-		// etc. Called on framework-owned controllers; receiver stripped.
-		// Also appears as a StatefulWidget lifecycle override body calling
-		// `controller.dispose()`. Dart-gate prevents collision with user
-		// `dispose()` methods in other ecosystems.
-		regexp.MustCompile(`^dispose$`),
-		// `clear` — TextEditingController.clear() / List.clear().
-		regexp.MustCompile(`^clear$`),
-	}
-
 	// dynamicPatternsByLang dispatches a normalized language tag to its
 	// per-language pattern slice. Keys are lower-case canonical names; the
 	// resolver normalizes incoming tags before lookup.
-	dynamicPatternsByLang = map[string][]*regexp.Regexp{"python":     pythonDynamicPatterns,
-		"go":         goDynamicPatterns,
-		"javascript": jsDynamicPatterns,
-		"typescript": jsDynamicPatterns,
-		"ruby":       rubyDynamicPatterns,
-		"java":       jvmDynamicPatterns,
-		"kotlin":     jvmDynamicPatterns,
-		"scala":      jvmDynamicPatterns,
-		"jvm":        jvmDynamicPatterns,
-		"hcl":        hclDynamicPatterns,
-		"terraform":  hclDynamicPatterns,
-		"csharp":     csharpDynamicPatterns,
-		"razor":      csharpDynamicPatterns,
-		"rust":       rustDynamicPatterns,
-		"lua":        luaDynamicPatterns,
-		"swift":      swiftDynamicPatterns,
-		"dart":       dartDynamicPatterns,
-		"elixir":     elixirDynamicPatterns,
-		"zig":        zigDynamicPatterns,
-	}
+	// The map is populated by init() functions in dynamic_patterns_<lang>.go
+	// files — one per language. Adding a new language only requires two new
+	// files; refs.go is untouched.
+	dynamicPatternsByLang = map[string][]*regexp.Regexp{}
 )
 
 // normalizeLang lowercases a language tag and maps a few common aliases to
